refactor(runtime): split registration_class.rs into cohesive submodules (§7-8)
Pure relocation: split the 3964-line registration_class.rs into flat
sibling modules under src/runtime/, each an `impl Interpreter` block with
`use super::*;`. No logic changes.

New layout:
- registration_class.rs (486): prelude (shared free fns, AttrValidationCtx,
  language_revision_letter) + handle/resolve helper methods
- registration_class_decl.rs (1815): register_class_decl (indivisible giant)
- registration_role.rs (874): role candidate helpers + register_role_decl
  (indivisible giant)
- registration_class_augment.rs (580): augment_class/role, subset, cunion, pun
- registration_class_attr.rs (227): attribute-declaration validation +
  store_language_revision_from_version

The two giant methods (register_class_decl ~1810 lines, register_role_decl
~665 lines) each remain alone in their own sibling file; they are not
internally divisible without changing logic.
