+++ src/runtime/mod.rs
@@ -265,6 +265,10 @@ mod regex_parse_ltm;
 mod regex_parse_modifier;
 mod registration;
 mod registration_class;
+mod registration_class_attr;
+mod registration_class_augment;
+mod registration_class_decl;
+mod registration_role;
 pub(crate) mod registration_sub;
 mod registry;
 pub(crate) mod resolution;
