+++ src/runtime/registration_class.rs
@@ -1,8 +1,7 @@
 use super::*;
-use crate::ast::{HandleSpec, ParamDef, PhaserKind};
-use crate::symbol::Symbol;
+use crate::ast::{HandleSpec, ParamDef};
 
-type ResolvedRoleCandidate = (RoleDef, Vec<String>, Vec<Value>);
+pub(super) type ResolvedRoleCandidate = (RoleDef, Vec<String>, Vec<Value>);
 
 /// Whether a built-in type is a concrete (non-composable) class that cannot be
 /// `does`-composed by a class — `class B does Int {}` is
@@ -11,14 +10,14 @@ type ResolvedRoleCandidate = (RoleDef, Vec<String>, Vec<Value>);
 /// (Real/Numeric/Positional/Baggy/Setty/Stringy/...) and composes fine. A
 /// denylist is used (rather than an allowlist of composable roles) so new or
 /// uncommon built-in roles like Baggy/Setty/Mixy are not wrongly rejected.
-fn is_non_composable_builtin(name: &str) -> bool {
+pub(super) fn is_non_composable_builtin(name: &str) -> bool {
     matches!(
         name,
         "Int" | "UInt" | "Str" | "Num" | "Rat" | "FatRat" | "Complex" | "Cool" | "Any" | "Mu"
     )
 }
 
-fn type_value_name(value: &Value) -> String {
+pub(super) fn type_value_name(value: &Value) -> String {
     match value {
         Value::Package(name) => name.resolve(),
         Value::ParametricRole {
@@ -41,7 +40,7 @@ fn type_value_name(value: &Value) -> String {
     }
 }
 
-fn builtin_role_def() -> RoleDef {
+pub(super) fn builtin_role_def() -> RoleDef {
     RoleDef {
         attributes: Vec::new(),
         methods: HashMap::new(),
@@ -59,7 +58,7 @@ fn builtin_role_def() -> RoleDef {
 }
 
 /// Intermediate representation for resolved handle specs.
-enum ResolvedHandle {
+pub(crate) enum ResolvedHandle {
     /// Forward `exposed` method to `target` method on the object in `attr_var`.
     Method {
         exposed: String,
@@ -73,7 +72,7 @@ enum ResolvedHandle {
 }
 
 /// Apply resolved handles to methods map and wildcard handles.
-fn apply_resolved_handles(
+pub(super) fn apply_resolved_handles(
     handles: &[ResolvedHandle],
     methods: &mut HashMap<String, Vec<MethodDef>>,
     wildcard_handles: &mut Vec<String>,
@@ -128,7 +127,7 @@ fn delegation_slurpy_param() -> ParamDef {
 
 /// Create a delegation MethodDef that forwards method calls to `target_method`
 /// on the object in `attr_var_name`.
-fn make_delegation_method(attr_var_name: &str, target_method: &str) -> MethodDef {
+pub(super) fn make_delegation_method(attr_var_name: &str, target_method: &str) -> MethodDef {
     MethodDef {
         params: vec!["@_".to_string(), "%_".to_string()],
         param_defs: vec![delegation_slurpy_param(), delegation_double_slurpy_param()],
@@ -184,11 +183,11 @@ pub(crate) struct ClassDeclModifiers<'a> {
     pub(crate) language_version: &'a str,
 }
 
-fn parse_role_type_args(input: &str) -> Vec<String> {
+pub(super) fn parse_role_type_args(input: &str) -> Vec<String> {
     split_balanced_comma_list(input)
 }
 
-fn looks_like_type_arg_expr(input: &str) -> bool {
+pub(super) fn looks_like_type_arg_expr(input: &str) -> bool {
     let trimmed = input.trim();
     if trimmed.is_empty() {
         return false;
@@ -202,7 +201,7 @@ fn looks_like_type_arg_expr(input: &str) -> bool {
     })
 }
 
-fn should_treat_role_arg_as_type_expr(input: &str) -> bool {
+pub(super) fn should_treat_role_arg_as_type_expr(input: &str) -> bool {
     let trimmed = input.trim();
     // Colonpair syntax like `:a(1)` or `:foo(42)` is a named argument, not a type.
     if trimmed.starts_with(':')
@@ -220,7 +219,7 @@ fn should_treat_role_arg_as_type_expr(input: &str) -> bool {
 /// Substitute type parameters in a method definition.
 /// E.g., if type_subs = [("T", "Str:D(Numeric)")], then any param with
 /// type_constraint "T" becomes "Str:D(Numeric)".
-fn substitute_type_params_in_method(
+pub(super) fn substitute_type_params_in_method(
     method: &MethodDef,
     type_subs: &[(String, String)],
 ) -> MethodDef {
@@ -288,11 +287,30 @@ fn substitute_type_params_in_method(
     }
 }
 
+/// Context threaded through the `$!attr` declaration validators so that an
+/// undeclared private attribute can be reported as a fully-populated
+/// `X::Attribute::Undeclared` (with `package-name`/`package-kind`).
+pub(crate) struct AttrValidationCtx<'a> {
+    pub(crate) attrs: &'a HashSet<String>,
+    pub(crate) pkg_name: &'a str,
+    pub(crate) pkg_kind: &'a str,
+}
+
+/// Extract the language revision letter from a version string like "6.c", "6.d", "6.e".
+pub(super) fn language_revision_letter(version: &str) -> String {
+    // Version is like "6.c", "6.d", "6.e" — extract the letter after the dot
+    if let Some(letter) = version.strip_prefix("6.") {
+        letter.chars().next().unwrap_or('c').to_string()
+    } else {
+        "c".to_string()
+    }
+}
+
 impl Interpreter {
     /// Apply `handles` specifications to a class definition.
     /// For type-based handles, collects method names from the referenced type
     /// first, then applies them without holding borrows on self.
-    fn apply_handle_specs(
+    pub(crate) fn apply_handle_specs(
         &self,
         specs: &[HandleSpec],
         attr_var_name: &str,
@@ -307,7 +325,7 @@ impl Interpreter {
     }
 
     /// Apply `handles` specifications to a role definition.
-    fn apply_handle_specs_to_role(
+    pub(crate) fn apply_handle_specs_to_role(
         &self,
         specs: &[HandleSpec],
         attr_var_name: &str,
@@ -323,7 +341,7 @@ impl Interpreter {
 
     /// Resolve handle specs to concrete (exposed_name, target_method, attr_var_name) tuples
     /// or wildcard/regex entries. This step only reads from self (immutable borrow).
-    fn resolve_handle_specs_to_names(
+    pub(crate) fn resolve_handle_specs_to_names(
         &self,
         specs: &[HandleSpec],
         attr_var_name: &str,
@@ -403,206 +421,7 @@ impl Interpreter {
         names
     }
 
-    fn eval_role_arg_values(&mut self, arg_exprs: &[String]) -> Result<Vec<Value>, RuntimeError> {
-        let mut values = Vec::with_capacity(arg_exprs.len());
-        for expr in arg_exprs {
-            if expr.trim_start().starts_with("::") {
-                return Err(RuntimeError::new(
-                    "X::Syntax::Malformed: cannot use ::T in role application".to_string(),
-                ));
-            }
-            if should_treat_role_arg_as_type_expr(expr) {
-                values.push(Value::Package(Symbol::intern(expr.trim())));
-                continue;
-            }
-            match crate::parse_dispatch::parse_source(expr)
-                .and_then(|(stmts, _)| self.eval_block_value(&stmts))
-            {
-                Ok(value) => values.push(value),
-                Err(_) if looks_like_type_arg_expr(expr) => {
-                    values.push(Value::Package(Symbol::intern(expr.trim())));
-                }
-                Err(err) => return Err(err),
-            }
-        }
-        Ok(values)
-    }
-
-    fn role_constraint_specificity(&self, constraint: Option<&str>) -> i32 {
-        let Some(constraint) = constraint else {
-            return 0;
-        };
-        if constraint.starts_with("::") {
-            return 1;
-        }
-        if constraint == "Any" || constraint == "Mu" {
-            return 2;
-        }
-        if let Some(def) = self.registry().classes.get(constraint) {
-            return 10 + def.parents.len() as i32;
-        }
-        if self.registry().roles.contains_key(constraint) {
-            return 9;
-        }
-        8
-    }
-
-    fn role_candidate_specificity_score(&self, param_defs: &[ParamDef]) -> i32 {
-        let mut score = 0i32;
-        for pd in param_defs.iter().filter(|pd| !pd.named) {
-            score += self.role_constraint_specificity(pd.type_constraint.as_deref());
-            if pd.where_constraint.is_some() {
-                score += 20;
-            }
-            if pd.literal_value.is_some() {
-                score += 30;
-            }
-        }
-        score
-    }
-
-    fn role_candidate_arity_ok(&self, args: &[Value], param_defs: &[ParamDef]) -> bool {
-        if param_defs.is_empty() {
-            return args.is_empty();
-        }
-        let positional_arg_count = args
-            .iter()
-            .filter(|arg| !matches!(arg, Value::Pair(..)))
-            .count();
-        let positional_params: Vec<&ParamDef> = param_defs.iter().filter(|pd| !pd.named).collect();
-        let has_positional_slurpy = positional_params
-            .iter()
-            .any(|pd| pd.slurpy && !pd.name.starts_with('%'));
-        let required = positional_params
-            .iter()
-            .filter(|pd| !pd.slurpy && pd.default.is_none() && !pd.optional_marker)
-            .count();
-        if positional_arg_count < required {
-            return false;
-        }
-        if !has_positional_slurpy && positional_arg_count > positional_params.len() {
-            return false;
-        }
-        true
-    }
-
-    fn resolve_role_candidate(
-        &mut self,
-        parent: &str,
-    ) -> Result<Option<ResolvedRoleCandidate>, RuntimeError> {
-        let parent = self.resolve_declared_type_name(parent);
-        if let Some(bracket_start) = parent.find('[') {
-            let args_str = &parent[bracket_start + 1..parent.len() - 1];
-            let arg_exprs = parse_role_type_args(args_str);
-            if arg_exprs
-                .iter()
-                .any(|expr| expr.trim_start().starts_with("::"))
-            {
-                return Err(RuntimeError::new(
-                    "X::Syntax::Malformed: cannot use ::T in role application".to_string(),
-                ));
-            }
-        }
-
-        let base_role_name = if let Some(bracket) = parent.find('[') {
-            &parent[..bracket]
-        } else {
-            &parent
-        };
-        let Some(candidates) = self.registry().role_candidates.get(base_role_name).cloned() else {
-            if let Some(role) = self.registry().roles.get(base_role_name).cloned() {
-                return Ok(Some((role, Vec::new(), Vec::new())));
-            }
-            if matches!(base_role_name, "Positional" | "Associative" | "Callable") {
-                return Ok(Some((builtin_role_def(), Vec::new(), Vec::new())));
-            }
-            return Ok(None);
-        };
-
-        let arg_exprs = if let Some(bracket_start) = parent.find('[') {
-            let args_str = &parent[bracket_start + 1..parent.len() - 1];
-            parse_role_type_args(args_str)
-        } else {
-            Vec::new()
-        };
-        let arg_values = self.eval_role_arg_values(&arg_exprs)?;
-
-        let mut matches: Vec<(RoleCandidateDef, i32, usize)> = candidates
-            .into_iter()
-            .enumerate()
-            .filter_map(|(idx, candidate)| {
-                let candidate_param_names = candidate
-                    .type_param_defs
-                    .iter()
-                    .map(|pd| pd.name.clone())
-                    .collect::<Vec<_>>();
-                let ok = if self.role_candidate_arity_ok(&arg_values, &candidate.type_param_defs) {
-                    let saved_env = self.env.clone();
-                    let ok = self
-                        .bind_function_args_values(
-                            &candidate.type_param_defs,
-                            &candidate_param_names,
-                            &arg_values,
-                        )
-                        .is_ok();
-                    self.env = saved_env;
-                    ok
-                } else {
-                    false
-                };
-                if ok {
-                    Some((
-                        candidate.clone(),
-                        self.role_candidate_specificity_score(&candidate.type_param_defs),
-                        idx,
-                    ))
-                } else {
-                    None
-                }
-            })
-            .collect();
-
-        if matches.is_empty() {
-            return Err(RuntimeError::typed_msg(
-                "X::Role::Parametric::NoSuchCandidate",
-                "No matching candidate found for the parametric role",
-            ));
-        }
-
-        matches.sort_by(|a, b| b.1.cmp(&a.1).then(b.2.cmp(&a.2)));
-        let selected = matches.remove(0).0;
-        // Properly bind args (handling named params, defaults, etc.) and extract
-        // resolved values per param name, instead of using raw positional zip.
-        let resolved_values = if !selected.type_param_defs.is_empty() {
-            let saved_env = self.env.clone();
-            let candidate_param_names: Vec<String> = selected
-                .type_param_defs
-                .iter()
-                .map(|pd| pd.name.clone())
-                .collect();
-            let _ = self.bind_function_args_values(
-                &selected.type_param_defs,
-                &candidate_param_names,
-                &arg_values,
-            );
-            let mut resolved = Vec::with_capacity(selected.type_params.len());
-            for param_name in &selected.type_params {
-                let value = self.env.get(param_name).cloned().unwrap_or(Value::Nil);
-                resolved.push(value);
-            }
-            self.env = saved_env;
-            resolved
-        } else {
-            arg_values
-        };
-        Ok(Some((
-            selected.role_def,
-            selected.type_params,
-            resolved_values,
-        )))
-    }
-
-    fn resolve_declared_type_name(&self, name: &str) -> String {
+    pub(crate) fn resolve_declared_type_name(&self, name: &str) -> String {
         let (base, suffix) = if let Some(bracket) = name.find('[') {
             (&name[..bracket], &name[bracket..])
         } else {
@@ -663,3302 +482,5 @@ impl Interpreter {
             }
         }
         name.to_string()
-    }
-
-    /// Register a class declaration. Returns a list of unknown lowercase parent
-    /// names that were deferred for custom `trait_mod:<is>` dispatch (empty if
-    /// no custom traits are needed).
-    pub(crate) fn register_class_decl(
-        &mut self,
-        name: &str,
-        parents: &[String],
-        modifiers: ClassDeclModifiers<'_>,
-        body: &[Stmt],
-    ) -> Result<Vec<String>, RuntimeError> {
-        self.clear_private_zeroarg_method_cache();
-        let ClassDeclModifiers {
-            class_is_rw,
-            is_hidden,
-            is_lexical,
-            hidden_parents,
-            does_parents,
-            language_version: class_language_version,
-        } = modifiers;
-        let class_lang_rev = language_revision_letter(class_language_version);
-        // Normalize parent names: strip leading `::` (indirect name lookup syntax).
-        // `is ::Foo` means the same as `is Foo` in Raku.
-        let strip_colons = |s: &str| s.strip_prefix("::").unwrap_or(s).to_string();
-        let parents: Vec<String> = parents.iter().map(|p| strip_colons(p)).collect();
-        let parents = parents.as_slice();
-        let does_parents: Vec<String> = does_parents.iter().map(|p| strip_colons(p)).collect();
-        let does_parents = does_parents.as_slice();
-        let hidden_parents: Vec<String> = hidden_parents.iter().map(|p| strip_colons(p)).collect();
-        let hidden_parents = hidden_parents.as_slice();
-        // Snapshot the previous registry state for this class under a single read
-        // guard (all values are owned/cloned, no re-entry) so a redefinition can be
-        // rolled back if the new body fails.
-        let (
-            prev_class,
-            prev_hidden,
-            prev_hidden_defer,
-            prev_composed_roles,
-            prev_role_param_bindings,
-        ) = {
-            let reg = self.registry();
-            (
-                reg.classes.get(name).cloned(),
-                reg.hidden_classes.contains(name),
-                reg.hidden_defer_parents.get(name).cloned(),
-                reg.class_composed_roles.get(name).cloned(),
-                reg.class_role_param_bindings.get(name).cloned(),
-            )
-        };
-        // Clear `is Type` trait entries for this class (they'll be re-populated from the body).
-        self.registry_mut()
-            .class_attribute_is_types
-            .retain(|(cn, _), _| cn != name);
-
-        // Rollback writes are purely registry mutations with no user-code
-        // re-entry, so they take a single write guard for the whole block.
-        let restore_previous_state = |this: &mut Self| {
-            let mut reg = this.registry_mut();
-            if let Some(class_def) = prev_class.clone() {
-                reg.classes.insert(name.to_string(), class_def);
-            } else {
-                reg.classes.remove(name);
-            }
-            if prev_hidden {
-                reg.hidden_classes.insert(name.to_string());
-            } else {
-                reg.hidden_classes.remove(name);
-            }
-            if let Some(hidden) = prev_hidden_defer.clone() {
-                reg.hidden_defer_parents.insert(name.to_string(), hidden);
-            } else {
-                reg.hidden_defer_parents.remove(name);
-            }
-            if let Some(composed) = prev_composed_roles.clone() {
-                reg.class_composed_roles.insert(name.to_string(), composed);
-            } else {
-                reg.class_composed_roles.remove(name);
-            }
-            if let Some(bindings) = prev_role_param_bindings.clone() {
-                reg.class_role_param_bindings
-                    .insert(name.to_string(), bindings);
-            } else {
-                reg.class_role_param_bindings.remove(name);
-            }
-        };
-
-        // Detect X::Redeclaration when a class redefines a role in the same scope.
-        // Only check user-declared roles (not pre-registered builtins like Iterator).
-        // Lexical classes (`my class`) are allowed to shadow outer role names.
-        if !is_lexical && self.registry().user_declared_roles.contains(name) {
-            // Only report redeclaration for non-stub class bodies
-            let body_non_stub: Vec<_> = body
-                .iter()
-                .filter(|s| !matches!(s, Stmt::SetLine(_)))
-                .collect();
-            let class_is_stub = body_non_stub.len() == 1
-                && matches!(body_non_stub[0], Stmt::Expr(Expr::Call { name: fn_name, .. })
-                    if *fn_name == "__mutsu_stub_die" || *fn_name == "__mutsu_stub_warn");
-            if !class_is_stub {
-                return Err(RuntimeError::redeclaration("symbol", name));
-            }
-        }
-
-        // Detect stub body: `class Foo { ... }` — body is a stub operator call
-        // Filter SetLine annotations which don't affect the stub nature.
-        let body_no_setline: Vec<_> = body
-            .iter()
-            .filter(|s| !matches!(s, Stmt::SetLine(_)))
-            .collect();
-        let is_stub_body = body_no_setline.len() == 1
-            && matches!(body_no_setline[0], Stmt::Expr(Expr::Call { name: fn_name, .. })
-                if *fn_name == "__mutsu_stub_die" || *fn_name == "__mutsu_stub_warn");
-
-        // If this is a stub registration but the class already exists and is
-        // NOT a stub (i.e., it was already filled in by a hoisted real
-        // declaration), skip the stub registration to avoid overwriting the
-        // real class definition.
-        if is_stub_body
-            && self.registry().classes.contains_key(name)
-            && !self.registry().class_stubs.contains(name)
-        {
-            return Ok(Vec::new());
-        }
-
-        // Validate that all parent classes exist
-        // Allow inheriting from built-in types that may not be in the classes HashMap
-        const BUILTIN_TYPES: &[&str] = &[
-            "Mu",
-            "Any",
-            "Cool",
-            "Int",
-            "Num",
-            "Str",
-            "Bool",
-            "Rat",
-            "FatRat",
-            "Complex",
-            "Array",
-            "Hash",
-            "List",
-            "Map",
-            "Set",
-            "Bag",
-            "Mix",
-            "SetHash",
-            "BagHash",
-            "MixHash",
-            "Range",
-            "Pair",
-            "IO",
-            "IO::Path",
-            "IO::Handle",
-            "IO::Spec",
-            "IO::Spec::Unix",
-            "IO::Spec::Win32",
-            "IO::Spec::Cygwin",
-            "IO::Spec::QNX",
-            "Regex",
-            "Match",
-            "Junction",
-            "Exception",
-            "Failure",
-            "Version",
-            "Nil",
-            "Block",
-            "Code",
-            "Routine",
-            "Sub",
-            "Method",
-            "Seq",
-            "Slip",
-            "Whatever",
-            "WhateverCode",
-            "HyperWhatever",
-            "Callable",
-            "Numeric",
-            "Real",
-            "Stringy",
-            "Positional",
-            "Associative",
-            "Order",
-            "Endian",
-            "Proc",
-            "Proc::Async",
-            "Supply",
-            "Supplier",
-            "Setty",
-            "Baggy",
-            "Mixy",
-            "Date",
-            "DateTime",
-            "Capture",
-            "Grammar",
-            "Parameter",
-            "Proxy",
-            "Signature",
-            "Stash",
-            "Metamodel::ClassHOW",
-            "Perl6::Metamodel::ClassHOW",
-        ];
-        let mut deferred_custom_traits: Vec<String> = Vec::new();
-        for parent in parents {
-            let resolved_parent_name = self.resolve_declared_type_name(parent);
-            // Strip type arguments for validation (e.g., "R[Str:D(Numeric)]" -> "R")
-            let base_parent = if let Some(bracket) = resolved_parent_name.find('[') {
-                &resolved_parent_name[..bracket]
-            } else {
-                resolved_parent_name.as_str()
-            };
-            // Strip leading `::` for comparison (e.g., `is ::F` refers to `F`)
-            let resolved_parent = base_parent.strip_prefix("::").unwrap_or(base_parent);
-            if resolved_parent == name {
-                return Err(RuntimeError::new(format!(
-                    "X::Inheritance::SelfInherit: class '{}' cannot inherit from itself",
-                    name
-                )));
-            }
-            if !self.registry().classes.contains_key(base_parent)
-                && !BUILTIN_TYPES.contains(&base_parent)
-                && !self.registry().roles.contains_key(base_parent)
-                && !self.registry().enum_types.contains_key(base_parent)
-            {
-                // Use X::InvalidType for `does`/`hides` parents,
-                // X::Inheritance::UnknownParent for `is` parents.
-                if does_parents.contains(parent) || hidden_parents.contains(parent) {
-                    return Err(RuntimeError::new(format!(
-                        "X::InvalidType: Invalid typename '{}'",
-                        resolved_parent_name
-                    )));
-                }
-                // If trait_mod:<is> is defined, defer unknown lowercase parents
-                // to custom trait dispatch instead of erroring.
-                if (self.has_proto("trait_mod:<is>") || self.has_multi_candidates("trait_mod:<is>"))
-                    && resolved_parent
-                        .chars()
-                        .next()
-                        .is_some_and(|c| c.is_ascii_lowercase())
-                {
-                    deferred_custom_traits.push(resolved_parent_name.to_string());
-                    continue;
-                }
-                // A name that is declared as a `package` (or module) exists but
-                // does not support inheritance: `package A {}; class B is A {}`
-                // is X::Inheritance::Unsupported, not an unknown-parent error.
-                if self.chain_declared_packages.contains(base_parent)
-                    || self
-                        .chain_declared_packages
-                        .contains(resolved_parent_name.as_str())
-                {
-                    let msg = format!(
-                        "{} does not support inheritance, so {} cannot inherit from it",
-                        resolved_parent_name, name
-                    );
-                    let mut attrs = HashMap::new();
-                    attrs.insert("child-typename".to_string(), Value::str(name.to_string()));
-                    attrs.insert(
-                        "parent".to_string(),
-                        Value::Package(crate::symbol::Symbol::intern(
-                            resolved_parent_name.as_str(),
-                        )),
-                    );
-                    attrs.insert("message".to_string(), Value::str(msg));
-                    return Err(RuntimeError::typed("X::Inheritance::Unsupported", attrs));
-                }
-                {
-                    // Suggest close known type names (Did-you-mean).
-                    let suggestions = self.suggest_type_names(resolved_parent_name.as_str());
-                    let mut msg = format!(
-                        "'{}' cannot inherit from '{}' because it is unknown.",
-                        name, resolved_parent_name
-                    );
-                    if suggestions.len() == 1 {
-                        msg.push_str(&format!("\nDid you mean '{}'?", suggestions[0]));
-                    } else if suggestions.len() > 1 {
-                        msg.push_str("\nDid you mean one of these?\n");
-                        for s in &suggestions {
-                            msg.push_str(&format!("    '{}'\n", s));
-                        }
-                    }
-                    let mut attrs = HashMap::new();
-                    attrs.insert("child-name".to_string(), Value::str(name.to_string()));
-                    attrs.insert("child".to_string(), Value::str(name.to_string()));
-                    attrs.insert(
-                        "parent-name".to_string(),
-                        Value::str(resolved_parent_name.to_string()),
-                    );
-                    attrs.insert(
-                        "parent".to_string(),
-                        Value::str(resolved_parent_name.to_string()),
-                    );
-                    attrs.insert(
-                        "suggestions".to_string(),
-                        Value::array(suggestions.into_iter().map(Value::str).collect()),
-                    );
-                    attrs.insert("message".to_string(), Value::str(msg));
-                    return Err(RuntimeError::typed("X::Inheritance::UnknownParent", attrs));
-                }
-            }
-            // A `does` target that is a non-composable built-in concrete class
-            // (Int, Str, Num, Cool, Any, Mu, ...) — as opposed to a composable
-            // built-in role (Real, Numeric, Positional, Iterable, ...) — raises
-            // X::Composition::NotComposable.
-            if does_parents.contains(parent)
-                && !self.registry().roles.contains_key(resolved_parent)
-                && BUILTIN_TYPES.contains(&resolved_parent)
-                && is_non_composable_builtin(resolved_parent)
-            {
-                let msg = format!(
-                    "{} is not composable, so {} cannot compose it",
-                    resolved_parent, name
-                );
-                let mut attrs = HashMap::new();
-                attrs.insert("target-name".to_string(), Value::str(name.to_string()));
-                attrs.insert(
-                    "composer".to_string(),
-                    Value::Package(crate::symbol::Symbol::intern(resolved_parent)),
-                );
-                attrs.insert("message".to_string(), Value::str(msg.clone()));
-                let ex = Value::make_instance(
-                    crate::symbol::Symbol::intern("X::Composition::NotComposable"),
-                    attrs,
-                );
-                let mut err = RuntimeError::new(&msg);
-                err.exception = Some(Box::new(ex));
-                return Err(err);
-            }
-            // Check that `does` targets are actually roles, not classes
-            if does_parents.contains(parent)
-                && self.registry().classes.contains_key(resolved_parent)
-                && !self.registry().roles.contains_key(resolved_parent)
-                && !BUILTIN_TYPES.contains(&resolved_parent)
-            {
-                return Err(RuntimeError::new(format!(
-                    "'{}' cannot compose '{}' because it is not a role",
-                    name, resolved_parent
-                )));
-            }
-            // Check if parent is a stub (not yet composed)
-            if self.registry().class_stubs.contains(resolved_parent) {
-                let message = format!(
-                    "'{}' cannot inherit from '{}' because '{}' isn't composed yet (maybe it is stubbed)",
-                    name, resolved_parent, resolved_parent
-                );
-                let mut attrs = HashMap::new();
-                attrs.insert("child-name".to_string(), Value::str(name.to_string()));
-                attrs.insert(
-                    "parent-name".to_string(),
-                    Value::str(resolved_parent.to_string()),
-                );
-                attrs.insert("message".to_string(), Value::str(message.clone()));
-                let ex = Value::make_instance(
-                    crate::symbol::Symbol::intern("X::Inheritance::NotComposed"),
-                    attrs,
-                );
-                let mut err =
-                    RuntimeError::new(format!("X::Inheritance::NotComposed: {}", message));
-                err.exception = Some(Box::new(ex));
-                return Err(err);
-            }
-        }
-        let mut class_def = ClassDef {
-            parents: parents.to_vec(),
-            attributes: Vec::new(),
-            attribute_types: HashMap::new(),
-            attribute_smileys: HashMap::new(),
-            attribute_built: HashMap::new(),
-            methods: HashMap::new(),
-            native_methods: HashSet::new(),
-            mro: Vec::new(),
-            wildcard_handles: Vec::new(),
-            alias_attributes: HashSet::new(),
-            class_level_attrs: HashMap::new(),
-        };
-        if is_hidden {
-            self.registry_mut().hidden_classes.insert(name.to_string());
-        } else {
-            self.registry_mut().hidden_classes.remove(name);
-        }
-        if hidden_parents.is_empty() {
-            self.registry_mut().hidden_defer_parents.remove(name);
-        } else {
-            self.registry_mut()
-                .hidden_defer_parents
-                .insert(name.to_string(), hidden_parents.iter().cloned().collect());
-        }
-        // Compose roles listed in the parents (from "does Role" or "is Role" in class header)
-        let mut composed_roles_list = Vec::new();
-        let mut punned_roles = Vec::new();
-        let mut hidden_punned_role_bases: HashSet<String> = HashSet::new();
-        let mut class_role_param_bindings: HashMap<String, Value> = HashMap::new();
-        for parent in parents {
-            let resolved_parent_name = self.resolve_declared_type_name(parent);
-            let base_role_name = resolved_parent_name
-                .split_once('[')
-                .map(|(b, _)| b)
-                .unwrap_or(resolved_parent_name.as_str());
-            if let Some((role, role_param_names, role_arg_values)) =
-                self.resolve_role_candidate(&resolved_parent_name)?
-            {
-                if role.is_stub_role {
-                    return Err(RuntimeError::typed_msg(
-                        "X::Role::Parametric::NoSuchCandidate",
-                        "No matching candidate found for the parametric role",
-                    ));
-                }
-                // Check for attribute conflicts detected during role composition
-                if let Some((attr_name, role_a, role_b)) = role.attribute_conflicts.first() {
-                    return Err(RuntimeError::new(format!(
-                        "Attribute '$!{}' conflicts in role '{}' composition: declared in both '{}' and '{}'",
-                        attr_name, base_role_name, role_a, role_b
-                    )));
-                }
-                // Check if this role was specified via `is` (punning) vs `does` (composition)
-                let is_punned = !does_parents.contains(parent);
-                if is_punned {
-                    punned_roles.push(resolved_parent_name.clone());
-                    if role.is_hidden {
-                        hidden_punned_role_bases.insert(base_role_name.to_string());
-                    }
-                }
-                composed_roles_list.push(resolved_parent_name.clone());
-                // Look up the role's language revision for submethod composition rules.
-                let role_lang_rev = self
-                    .type_metadata
-                    .get(base_role_name)
-                    .and_then(|m| m.get("language-revision"))
-                    .map(|v| v.to_string_value())
-                    .unwrap_or_else(|| "c".to_string());
-                // Submethods from roles are only composed when the class is 6.c
-                // AND the role is also 6.c. In 6.d+, submethods are never composed
-                // from roles.
-                let compose_submethods = class_lang_rev == "c" && role_lang_rev == "c";
-                // Collect type parameter substitutions for method type constraints.
-                let type_subs: Vec<(String, String)> = role_param_names
-                    .iter()
-                    .zip(role_arg_values.iter())
-                    .map(|(p, v)| (p.clone(), type_value_name(v)))
-                    .collect();
-                for (p, v) in role_param_names.iter().zip(role_arg_values.iter()) {
-                    class_role_param_bindings.insert(p.clone(), v.clone());
-                }
-                for attr in &role.attributes {
-                    if !class_def.attributes.iter().any(|(n, ..)| n == &attr.0) {
-                        class_def.attributes.push(attr.clone());
-                    }
-                }
-                for (mname, overloads) in &role.methods {
-                    // Skip methods declared with `my` scope -- they are role-private
-                    // and should not be composed into consuming classes.
-                    // Submethods (is_submethod=true) ARE composed only when both
-                    // the class and role share 6.c language revision.
-                    let non_my_overloads: Vec<&MethodDef> = overloads
-                        .iter()
-                        .filter(|md| !md.is_my || (md.is_submethod && compose_submethods))
-                        .collect();
-                    if non_my_overloads.is_empty() {
-                        continue;
-                    }
-                    let composed: Vec<MethodDef> = if type_subs.is_empty() {
-                        non_my_overloads
-                            .into_iter()
-                            .map(|md| {
-                                let mut method = md.clone();
-                                if method.original_role.is_none() {
-                                    method.original_role = method.role_origin.clone();
-                                }
-                                method.role_origin = Some(base_role_name.to_string());
-                                method
-                            })
-                            .collect()
-                    } else {
-                        non_my_overloads
-                            .into_iter()
-                            .map(|md| {
-                                let mut method = substitute_type_params_in_method(md, &type_subs);
-                                if method.original_role.is_none() {
-                                    method.original_role = method.role_origin.clone();
-                                }
-                                method.role_origin = Some(base_role_name.to_string());
-                                method
-                            })
-                            .collect()
-                    };
-                    class_def
-                        .methods
-                        .entry(mname.clone())
-                        .or_default()
-                        .extend(composed);
-                }
-                // Transfer wildcard handles from role to class
-                for wh in &role.wildcard_handles {
-                    if !class_def.wildcard_handles.contains(wh) {
-                        class_def.wildcard_handles.push(wh.clone());
-                    }
-                }
-                let role_param_values: HashMap<String, Value> = role_param_names
-                    .iter()
-                    .cloned()
-                    .zip(role_arg_values.iter().cloned())
-                    .collect();
-                // Execute deferred body statements from parameterized roles
-                // with concrete type parameter bindings. These statements
-                // (e.g., `my T $v .= new;`) may create closure variables that
-                // are referenced by composed methods, so we must keep their
-                // effects on the env (only clean up the type capture markers).
-                if !role.deferred_body_stmts.is_empty() {
-                    // Bind type parameters as type captures
-                    for (param_name, param_value) in &role_param_values {
-                        self.bind_type_capture(param_name, param_value);
-                    }
-                    for stmt in &role.deferred_body_stmts {
-                        self.run_block_raw(std::slice::from_ref(stmt))?;
-                    }
-                    // Remove type capture markers (but keep the variables
-                    // created by the deferred stmts for method closures)
-                    for param_name in role_param_values.keys() {
-                        self.env.remove(&format!("__type_capture__{}", param_name));
-                        // Don't remove the param name itself - methods may need it
-                    }
-                }
-                if let Some(parent_specs) =
-                    self.registry().role_parents.get(base_role_name).cloned()
-                {
-                    for parent_spec in parent_specs {
-                        let resolved_parent = if let Some(v) = role_param_values.get(&parent_spec) {
-                            type_value_name(v)
-                        } else if let Some((pbase, _)) = parent_spec.split_once('[') {
-                            let p_args_str = &parent_spec[pbase.len() + 1..parent_spec.len() - 1];
-                            let p_args = parse_role_type_args(p_args_str)
-                                .into_iter()
-                                .map(|arg| {
-                                    role_param_values
-                                        .get(&arg)
-                                        .map(|v| match v {
-                                            Value::Package(name) => name.resolve(),
-                                            other => other
-                                                .to_string_value()
-                                                .trim_start_matches('(')
-                                                .trim_end_matches(')')
-                                                .to_string(),
-                                        })
-                                        .unwrap_or(arg)
-                                })
-                                .collect::<Vec<_>>();
-                            format!("{pbase}[{}]", p_args.join(","))
-                        } else {
-                            parent_spec.clone()
-                        };
-                        let parent_base = resolved_parent
-                            .split_once('[')
-                            .map(|(b, _)| b)
-                            .unwrap_or(resolved_parent.as_str());
-                        if let Some(parent_role) = self.registry().roles.get(parent_base).cloned() {
-                            if !composed_roles_list.contains(&resolved_parent) {
-                                composed_roles_list.push(resolved_parent.clone());
-                            }
-                            let parent_type_subs: Vec<(String, String)> = if let Some(parent_tps) =
-                                self.registry().role_type_params.get(parent_base)
-                            {
-                                if let Some(bracket_start) = resolved_parent.find('[') {
-                                    let args_str = &resolved_parent
-                                        [bracket_start + 1..resolved_parent.len() - 1];
-                                    let args = parse_role_type_args(args_str);
-                                    parent_tps
-                                        .iter()
-                                        .zip(args.iter())
-                                        .map(|(p, a)| (p.clone(), a.clone()))
-                                        .collect()
-                                } else {
-                                    Vec::new()
-                                }
-                            } else {
-                                Vec::new()
-                            };
-                            for attr in &parent_role.attributes {
-                                if !class_def.attributes.iter().any(|(n, ..)| n == &attr.0) {
-                                    class_def.attributes.push(attr.clone());
-                                }
-                            }
-                            for (mname, overloads) in &parent_role.methods {
-                                // Skip methods declared with `my` scope -- role-private
-                                let non_my_overloads: Vec<&MethodDef> =
-                                    overloads.iter().filter(|md| !md.is_my).collect();
-                                if non_my_overloads.is_empty() {
-                                    continue;
-                                }
-                                // If the composing role (base_role_name, e.g. R2) defines
-                                // this method itself, it has already resolved the same-named
-                                // method it inherits from its parent role (parent_base, e.g.
-                                // R1). Do not re-propagate the parent's copy into the consumer
-                                // as an independent candidate -- doing so would create a spurious
-                                // X::Role::Composition::Conflict. The parent role's method is
-                                // still reachable via a qualified call (self.R1::method).
-                                let resolved_by_composing_role =
-                                    role.methods.get(mname).is_some_and(|defs| {
-                                        defs.iter().any(|d| d.role_origin.is_none() && !d.is_my)
-                                    });
-                                if resolved_by_composing_role {
-                                    continue;
-                                }
-                                let composed: Vec<MethodDef> = if parent_type_subs.is_empty() {
-                                    non_my_overloads
-                                        .into_iter()
-                                        .map(|md| {
-                                            let mut method = md.clone();
-                                            if method.original_role.is_none() {
-                                                method.original_role = method.role_origin.clone();
-                                            }
-                                            method.role_origin = Some(parent_base.to_string());
-                                            method
-                                        })
-                                        .collect()
-                                } else {
-                                    non_my_overloads
-                                        .into_iter()
-                                        .map(|md| {
-                                            let mut method = substitute_type_params_in_method(
-                                                md,
-                                                &parent_type_subs,
-                                            );
-                                            if method.original_role.is_none() {
-                                                method.original_role = method.role_origin.clone();
-                                            }
-                                            method.role_origin = Some(parent_base.to_string());
-                                            method
-                                        })
-                                        .collect()
-                                };
-                                class_def
-                                    .methods
-                                    .entry(mname.clone())
-                                    .or_default()
-                                    .extend(composed);
-                            }
-                        } else if self.registry().classes.contains_key(parent_base)
-                            && !class_def.parents.iter().any(|p| p == &resolved_parent)
-                        {
-                            class_def.parents.push(resolved_parent.clone());
-                        }
-                    }
-                }
-            } else if does_parents.contains(parent)
-                && self.registry().enum_types.contains_key(base_role_name)
-            {
-                // Enum used as a role via `does`: record it for method dispatch
-                self.registry_mut()
-                    .class_enum_roles
-                    .entry(name.to_string())
-                    .or_default()
-                    .push(base_role_name.to_string());
-            } else if does_parents.contains(parent)
-                && BUILTIN_TYPES.contains(&base_role_name)
-                && !self.registry().roles.contains_key(base_role_name)
-                && !composed_roles_list.contains(&resolved_parent_name)
-            {
-                // Built-in type used as a role via `does` (e.g., `does Numeric`,
-                // `does Real`): record in composed_roles_list so that role-based
-                // method dispatch (e.g., .Numeric on type objects) works correctly.
-                composed_roles_list.push(resolved_parent_name.clone());
-            }
-        }
-        if class_role_param_bindings.is_empty() {
-            self.registry_mut().class_role_param_bindings.remove(name);
-        } else {
-            self.registry_mut()
-                .class_role_param_bindings
-                .insert(name.to_string(), class_role_param_bindings);
-        }
-        // Handle role punning: `is Role` creates a punned class from the role
-        for punned_role in &punned_roles {
-            let base_role = punned_role
-                .split_once('[')
-                .map(|(b, _)| b)
-                .unwrap_or(punned_role.as_str());
-            // Create a punned class entry if one doesn't already exist
-            if !self.registry().classes.contains_key(punned_role.as_str())
-                && !self.registry().classes.contains_key(base_role)
-            {
-                // Collect class parents and composed roles recursively from role hierarchy
-                let mut punned_class_parents = Vec::new();
-                let mut punned_composed_roles = Vec::new();
-                let mut role_stack = vec![base_role.to_string()];
-                let mut seen_roles = HashSet::new();
-                while let Some(role_name) = role_stack.pop() {
-                    if !seen_roles.insert(role_name.clone()) {
-                        continue;
-                    }
-                    if let Some(rparents) = self.registry().role_parents.get(&role_name).cloned() {
-                        for rp in rparents {
-                            let rp_base = rp.split_once('[').map(|(b, _)| b).unwrap_or(rp.as_str());
-                            if self.registry().roles.contains_key(rp_base) {
-                                // It's a role - add as composed role and recurse
-                                if !punned_composed_roles.contains(&rp) {
-                                    punned_composed_roles.push(rp.clone());
-                                }
-                                role_stack.push(rp_base.to_string());
-                            } else if self.registry().classes.contains_key(rp_base)
-                                && !punned_class_parents.contains(&rp)
-                            {
-                                // It's a class - add as parent
-                                punned_class_parents.push(rp);
-                            }
-                        }
-                    }
-                }
-                let punned_class = ClassDef {
-                    parents: punned_class_parents,
-                    attributes: Vec::new(),
-                    attribute_types: HashMap::new(),
-                    attribute_smileys: HashMap::new(),
-                    attribute_built: HashMap::new(),
-                    methods: HashMap::new(),
-                    native_methods: HashSet::new(),
-                    mro: Vec::new(),
-                    wildcard_handles: Vec::new(),
-                    alias_attributes: HashSet::new(),
-                    class_level_attrs: HashMap::new(),
-                };
-                self.registry_mut()
-                    .classes
-                    .insert(base_role.to_string(), punned_class);
-                if !punned_composed_roles.is_empty() {
-                    self.registry_mut()
-                        .class_composed_roles
-                        .insert(base_role.to_string(), punned_composed_roles);
-                }
-                // Propagate hidden status from role to punned class
-                // Read the flag under the guard, drop it, then write (read->write on
-                // the same lock would deadlock).
-                let base_role_is_hidden = self
-                    .registry()
-                    .roles
-                    .get(base_role)
-                    .is_some_and(|role_def| role_def.is_hidden);
-                if base_role_is_hidden {
-                    self.registry_mut()
-                        .hidden_classes
-                        .insert(base_role.to_string());
-                }
-                if hidden_punned_role_bases.contains(base_role) {
-                    self.registry_mut()
-                        .hidden_classes
-                        .insert(base_role.to_string());
-                }
-                // Recompute MRO for the punned class
-                let mro = self.class_mro(base_role);
-                if let Some(cd) = self.registry_mut().classes.get_mut(base_role) {
-                    cd.mro = mro;
-                }
-            }
-            if hidden_punned_role_bases.contains(base_role) {
-                self.registry_mut()
-                    .hidden_classes
-                    .insert(base_role.to_string());
-            }
-        }
-        // Clear stale composed roles from previous registration
-        self.registry_mut().class_composed_roles.remove(name);
-        if !composed_roles_list.is_empty() {
-            // Propagate role parent classes to the class (recursively through sub-roles)
-            // When a role `R is C1` is composed into a class, C1 becomes a parent
-            {
-                let mut role_stack: Vec<String> = composed_roles_list
-                    .iter()
-                    .map(|r| {
-                        r.split_once('[')
-                            .map(|(b, _)| b)
-                            .unwrap_or(r.as_str())
-                            .to_string()
-                    })
-                    .collect();
-                let mut seen_roles = HashSet::new();
-                while let Some(role_name) = role_stack.pop() {
-                    if !seen_roles.insert(role_name.clone()) {
-                        continue;
-                    }
-                    if let Some(rparents) = self.registry().role_parents.get(&role_name).cloned() {
-                        for rp in rparents {
-                            let rp_base = rp.split_once('[').map(|(b, _)| b).unwrap_or(rp.as_str());
-                            if self.registry().roles.contains_key(rp_base) {
-                                // It's a sub-role, recurse
-                                role_stack.push(rp_base.to_string());
-                            } else if self.registry().classes.contains_key(rp_base)
-                                && !class_def.parents.contains(&rp)
-                            {
-                                class_def.parents.push(rp);
-                            }
-                        }
-                    }
-                }
-            }
-            self.registry_mut()
-                .class_composed_roles
-                .insert(name.to_string(), composed_roles_list.clone());
-            // Propagate `hides` from composed roles (and sub-roles) to the class
-            {
-                let mut role_stack: Vec<String> = composed_roles_list
-                    .iter()
-                    .map(|r| {
-                        r.split_once('[')
-                            .map(|(b, _)| b)
-                            .unwrap_or(r.as_str())
-                            .to_string()
-                    })
-                    .collect();
-                let mut seen_roles = HashSet::new();
-                while let Some(role_name) = role_stack.pop() {
-                    if !seen_roles.insert(role_name.clone()) {
-                        continue;
-                    }
-                    // Hoist the clone to a `let` so the read guard drops before the
-                    // registry_mut write below (read->write on the same lock deadlocks).
-                    let hides_list = self.registry().role_hides.get(&role_name).cloned();
-                    if let Some(hides_list) = hides_list {
-                        for hidden in hides_list {
-                            self.registry_mut()
-                                .hidden_defer_parents
-                                .entry(name.to_string())
-                                .or_default()
-                                .insert(hidden);
-                        }
-                    }
-                    // Recurse into sub-roles
-                    if let Some(rparents) = self.registry().role_parents.get(&role_name).cloned() {
-                        for rp in rparents {
-                            let rp_base = rp.split_once('[').map(|(b, _)| b).unwrap_or(rp.as_str());
-                            if self.registry().roles.contains_key(rp_base) {
-                                role_stack.push(rp_base.to_string());
-                            }
-                        }
-                    }
-                }
-            }
-        }
-        for stmt in body {
-            if let Stmt::TrustsDecl {
-                name: trusted_class,
-            } = stmt
-            {
-                self.registry_mut()
-                    .class_trusts
-                    .entry(name.to_string())
-                    .or_default()
-                    .insert(trusted_class.resolve());
-            }
-        }
-        // Make the class visible while its body executes so introspection calls
-        // like `A.^add_method(...)` inside the declaration can resolve `A`.
-        // Clear stale method wrap chains from a previous class with the same name.
-        self.method_wrap_chains.retain(|(cls, _, _), _| cls != name);
-        self.registry_mut()
-            .classes
-            .insert(name.to_string(), class_def.clone());
-        if is_stub_body {
-            self.registry_mut().class_stubs.insert(name.to_string());
-            self.registry_mut()
-                .classes
-                .insert(name.to_string(), class_def);
-            let mut stack = Vec::new();
-            let _ = self.compute_class_mro(name, &mut stack)?;
-            return Ok(deferred_custom_traits);
-        }
-        // Clear stub status now that the class has a real body (also clears
-        // package stub status for `package Foo { ... }; class Foo { }`).
-        {
-            let mut reg = self.registry_mut();
-            reg.class_stubs.remove(name);
-            reg.package_stubs.remove(name);
-        }
-        let saved_package = self.current_package();
-        let saved_env = self.env.clone();
-        self.set_current_package(name.to_string());
-        self.env
-            .insert("?CLASS".to_string(), Value::Package(Symbol::intern(name)));
-        // Flatten SyntheticBlock (from `has ($a, $b)` list form) so inner
-        // HasDecl statements are processed at the top level.
-        let flattened_body: Vec<&Stmt> = body
-            .iter()
-            .flat_map(|s| match s {
-                Stmt::SyntheticBlock(inner) => inner.iter().collect::<Vec<_>>(),
-                other => vec![other],
-            })
-            .collect();
-        // Pre-scan: collect attribute names declared directly in this class body.
-        // Combined with role-composed attributes already in class_def.attributes,
-        // this gives the full set of attributes valid for $!attr access.
-        let mut class_own_attrs: HashSet<String> = class_def
-            .attributes
-            .iter()
-            .map(|(n, ..)| n.clone())
-            .collect();
-        for stmt in &flattened_body {
-            if let Stmt::HasDecl {
-                name: attr_name,
-                is_our,
-                is_my,
-                ..
-            } = stmt
-                && !*is_our
-                && !*is_my
-            {
-                class_own_attrs.insert(attr_name.resolve());
-            }
-        }
-        let class_attr_ctx = AttrValidationCtx {
-            attrs: &class_own_attrs,
-            pkg_name: name,
-            pkg_kind: "class",
-        };
-        let saved_functions_keys: HashSet<String> = self
-            .registry()
-            .functions
-            .keys()
-            .map(|k| k.resolve())
-            .collect();
-        // LEAVE phasers declared at class-body scope (e.g. via
-        // `my $x will leave { ... }`) must fire when the class body is left,
-        // i.e. once all body statements have been processed. Collect them here
-        // and run them (LIFO) after the loop instead of executing them inline.
-        let mut class_leave_phasers: Vec<Vec<Stmt>> = Vec::new();
-        for stmt in flattened_body {
-            match stmt {
-                Stmt::Phaser {
-                    kind: PhaserKind::Leave,
-                    body,
-                } => {
-                    class_leave_phasers.push(body.clone());
-                }
-                Stmt::HasDecl {
-                    name: attr_name,
-                    is_public,
-                    default,
-                    handles,
-                    is_rw,
-                    is_readonly,
-                    type_constraint,
-                    type_smiley,
-                    is_required,
-                    sigil,
-                    where_constraint,
-                    is_alias,
-                    is_our,
-                    is_my,
-                    is_default,
-                    is_type,
-                    deprecated_message,
-                    is_built,
-                    unknown_traits,
-                } => {
-                    let attr_name_str = attr_name.resolve();
-
-                    // Handle unknown traits. If a user-defined `trait_mod:<is>`
-                    // (or `trait_mod:<will>`, etc.) can handle the trait, dispatch
-                    // to it with an Attribute introspection object; otherwise raise
-                    // X::Comp::Trait::Unknown. Kept in a separate method so its
-                    // locals don't inflate this already-large function's frame.
-                    if !unknown_traits.is_empty()
-                        && let Err(err) = self.apply_attribute_traits(
-                            unknown_traits,
-                            &attr_name_str,
-                            *sigil,
-                            *is_public,
-                            name,
-                            type_constraint.as_deref(),
-                        )
-                    {
-                        self.set_current_package(saved_package);
-                        self.env = saved_env;
-                        return Err(err);
-                    }
-
-                    // Handle class-level attributes (our $.x / my $.x)
-                    if *is_our || *is_my {
-                        // Evaluate the default value if present
-                        let initial_value = if let Some(expr) = default {
-                            self.eval_block_value(&[Stmt::Expr(expr.clone())])?
-                        } else {
-                            Value::Nil
-                        };
-                        class_def
-                            .class_level_attrs
-                            .insert(attr_name_str.clone(), initial_value);
-                        // Skip per-instance attribute registration
-                        continue;
-                    }
-
-                    // Check for duplicate attribute from role composition
-                    if class_def
-                        .attributes
-                        .iter()
-                        .any(|(n, ..)| n == &attr_name_str)
-                    {
-                        self.set_current_package(saved_package);
-                        self.env = saved_env;
-                        return Err(RuntimeError::new(format!(
-                            "X::Comp::Trait::Duplicate: attribute '{}' already exists in class '{}' (possibly from role composition)",
-                            attr_name_str, name,
-                        )));
-                    }
-                    let effective_is_rw = !*is_readonly && (*is_rw || (class_is_rw && *is_public));
-                    class_def.attributes.push((
-                        attr_name_str.clone(),
-                        *is_public,
-                        default.clone(),
-                        effective_is_rw,
-                        is_required.clone(),
-                        *sigil,
-                        where_constraint.as_ref().map(|wc| wc.as_ref().clone()),
-                    ));
-                    // Store `is default(...)` trait value for this attribute.
-                    // When is_default is set, the evaluated value is stored for
-                    // .VAR.default and Nil-restore behavior.
-                    // When only `default` is set (from `is default(X)` without `= value`),
-                    // also store it as the is_default trait value.
-                    if let Some(is_default_expr) = is_default {
-                        if let Ok(val) =
-                            self.eval_block_value(&[Stmt::Expr(is_default_expr.clone())])
-                        {
-                            // Type-check the default value against the attribute's type constraint
-                            if let Some(tc) = type_constraint {
-                                let type_ok = if matches!(val, Value::Nil) {
-                                    // Nil is only valid for untyped or Nil-accepting attributes
-                                    tc == "Any" || tc == "Mu" || tc.contains("Nil")
-                                } else {
-                                    self.type_matches_value(tc, &val)
-                                };
-                                if !type_ok {
-                                    let mut attrs = std::collections::HashMap::new();
-                                    attrs.insert(
-                                        "message".to_string(),
-                                        Value::str(format!(
-                                            "Type check failed in assignment to attribute; expected {} but got {}",
-                                            tc, super::utils::value_type_name(&val)
-                                        )),
-                                    );
-                                    attrs.insert(
-                                        "expected".to_string(),
-                                        Value::Package(crate::symbol::Symbol::intern(tc)),
-                                    );
-                                    attrs.insert(
-                                        "got".to_string(),
-                                        if matches!(val, Value::Nil) {
-                                            Value::Nil
-                                        } else {
-                                            val.clone()
-                                        },
-                                    );
-                                    let err = Value::make_instance(
-                                        crate::symbol::Symbol::intern(
-                                            "X::TypeCheck::Attribute::Default",
-                                        ),
-                                        attrs,
-                                    );
-                                    let mut runtime_err = RuntimeError::new(format!(
-                                        "X::TypeCheck::Attribute::Default: Type check failed for default value of attribute '{}'; expected {}, got {}",
-                                        attr_name_str,
-                                        tc,
-                                        super::utils::value_type_name(&val)
-                                    ));
-                                    runtime_err.exception = Some(Box::new(err));
-                                    self.set_current_package(saved_package);
-                                    self.env = saved_env;
-                                    return Err(runtime_err);
-                                }
-                            }
-                            self.registry_mut()
-                                .class_attribute_defaults
-                                .insert((name.to_string(), attr_name_str.clone()), val);
-                        }
-                    } else if default.is_some() {
-                        // No explicit `is default(X)`, but there IS a `default` expr.
-                        // This means either `has $.a = expr` or `has $.a is default(expr)` without `= value`.
-                        // We can't distinguish here, so we DON'T set class_attribute_defaults
-                        // (it would be wrong for `has $.a = 42` — Nil should give (Any), not 42).
-                    }
-                    if *is_alias {
-                        class_def.alias_attributes.insert(attr_name_str.clone());
-                    }
-                    if let Some(tc) = type_constraint {
-                        // Resolve ::?CLASS to the current class name
-                        let resolved_tc = tc.replace("::?CLASS", name);
-                        class_def
-                            .attribute_types
-                            .insert(attr_name_str.clone(), resolved_tc);
-                    }
-                    if let Some(ts) = type_smiley {
-                        class_def
-                            .attribute_smileys
-                            .insert(attr_name_str.clone(), ts.clone());
-                    }
-                    if let Some(built) = is_built {
-                        class_def
-                            .attribute_built
-                            .insert(attr_name_str.clone(), *built);
-                    }
-                    if let Some(it) = is_type {
-                        self.registry_mut()
-                            .class_attribute_is_types
-                            .insert((name.to_string(), attr_name_str.clone()), it.clone());
-                    }
-                    if let Some(dm) = deprecated_message {
-                        self.registry_mut()
-                            .class_attribute_deprecated
-                            .insert((name.to_string(), attr_name_str.clone()), dm.clone());
-                    }
-                    let attr_var_name = if *is_public {
-                        format!(".{}", attr_name_str)
-                    } else {
-                        format!("!{}", attr_name_str)
-                    };
-                    self.apply_handle_specs(handles, &attr_var_name, &mut class_def);
-                }
-                Stmt::MethodDecl {
-                    name: method_name,
-                    name_expr,
-                    params: _,
-                    param_defs,
-                    body: method_body,
-                    multi,
-                    is_rw,
-                    is_private,
-                    is_our,
-                    is_my,
-                    is_submethod,
-                    our_variable_form,
-                    return_type,
-                    is_default_candidate,
-                    deprecated_message,
-                    handles: method_handles,
-                    custom_traits: method_custom_traits,
-                    is_export: method_is_export,
-                    export_tags: method_export_tags,
-                } => {
-                    self.validate_private_access_in_stmts(name, method_body)?;
-                    Self::validate_attr_declared_in_class(&class_attr_ctx, method_body)?;
-                    // In BUILD/TWEAK submethods, :$!attr parameters must refer
-                    // to declared attributes; reject undeclared ones with
-                    // X::Attribute::Undeclared.
-                    {
-                        let mn = method_name.resolve();
-                        if mn == "BUILD" || mn == "TWEAK" {
-                            for pd in param_defs {
-                                if pd.name.starts_with('!') && pd.name != "!" {
-                                    let attr_name = &pd.name[1..]; // strip '!'
-                                    if !class_own_attrs.contains(attr_name) {
-                                        let err = Self::undeclared_attr_error(
-                                            &class_attr_ctx,
-                                            attr_name,
-                                            "!",
-                                        );
-                                        self.set_current_package(saved_package);
-                                        self.env = saved_env;
-                                        return Err(err);
-                                    }
-                                }
-                            }
-                        }
-                    }
-                    let resolved_method_name = if let Some(expr) = name_expr {
-                        self.eval_block_value(&[Stmt::Expr(expr.clone())])?
-                            .to_string_value()
-                    } else {
-                        method_name.resolve()
-                    };
-                    let mut effective_param_defs =
-                        Self::effective_method_param_defs(param_defs, is_hidden);
-                    // Auto-detect @_ usage in methods without explicit signatures
-                    if param_defs.is_empty() {
-                        let (use_positional, _) = Self::auto_signature_uses(method_body);
-                        if use_positional && !effective_param_defs.iter().any(|pd| pd.name == "@_")
-                        {
-                            // Insert @_ slurpy before the named %_ slurpy (if any)
-                            let insert_pos = effective_param_defs
-                                .iter()
-                                .position(|pd| pd.name.starts_with('%') && pd.slurpy)
-                                .unwrap_or(effective_param_defs.len());
-                            effective_param_defs.insert(
-                                insert_pos,
-                                ParamDef {
-                                    name: "@_".to_string(),
-                                    default: None,
-                                    multi_invocant: true,
-                                    required: false,
-                                    named: false,
-                                    slurpy: true,
-                                    double_slurpy: false,
-                                    onearg: false,
-                                    sigilless: false,
-                                    type_constraint: None,
-                                    literal_value: None,
-                                    sub_signature: None,
-                                    where_constraint: None,
-                                    traits: Vec::new(),
-                                    optional_marker: false,
-                                    outer_sub_signature: None,
-                                    code_signature: None,
-                                    is_invocant: false,
-                                    shape_constraints: None,
-                                },
-                            );
-                        }
-                    }
-                    let effective_params: Vec<String> = effective_param_defs
-                        .iter()
-                        .map(|p| p.name.clone())
-                        .collect();
-                    let def = MethodDef {
-                        params: effective_params.clone(),
-                        param_defs: effective_param_defs.clone(),
-                        body: std::sync::Arc::new(method_body.clone()),
-                        is_rw: *is_rw,
-                        is_private: *is_private,
-                        is_multi: *multi,
-                        // Use is_submethod for the MethodDef is_my flag, which
-                        // controls inheritance filtering (submethods not inherited).
-                        // `my method` and `our method` are NOT added to the method
-                        // table at all — they are only registered as functions.
-                        is_my: *is_submethod,
-                        role_origin: None,
-                        original_role: None,
-                        return_type: return_type.clone(),
-                        compiled_code: None,
-                        delegation: None,
-                        is_default: *is_default_candidate,
-                        deprecated_message: deprecated_message.clone(),
-                        is_submethod: *is_submethod,
-                    };
-                    // `my method` and `our method` are NOT part of the class
-                    // method table — they are only callable as functions.
-                    // Submethods (is_submethod=true) DO go in the table even
-                    // though they also have is_my=true from the parser.
-                    // `my method` and `our method` are NOT part of the class
-                    // method table — they are only callable as functions.
-                    // Submethods (is_submethod=true) DO go in the table even
-                    // though they also have is_my=true from the parser.
-                    // The `our &name = method name(...)` variable form
-                    // (our_variable_form=true) keeps the method in the table.
-                    let is_lexical_only = *is_my && !*is_submethod;
-                    let is_our_only = *is_our && !*our_variable_form;
-                    if !is_lexical_only && !is_our_only {
-                        if *multi {
-                            class_def
-                                .methods
-                                .entry(resolved_method_name.clone())
-                                .or_default()
-                                .push(def);
-                        } else {
-                            // Check for duplicate non-multi method definition.
-                            // Only error if the existing method was defined in
-                            // this class (not composed from a role) AND shares the
-                            // same privacy: a private `method !foo` and a public
-                            // `method foo` live in separate namespaces and do not
-                            // collide (they are stored together but dispatch filters
-                            // on `is_private`).
-                            let new_is_private = def.is_private;
-                            if let Some(existing) = class_def.methods.get(&resolved_method_name) {
-                                let conflicts = existing.iter().any(|m| {
-                                    m.role_origin.is_none() && m.is_private == new_is_private
-                                });
-                                if conflicts {
-                                    return Err(RuntimeError::new(format!(
-                                        "Package '{}' already has a method '{}' (did you mean to declare a multi method?)",
-                                        name, resolved_method_name
-                                    )));
-                                }
-                            }
-                            // A non-multi method replaces prior same-privacy
-                            // candidates but must preserve methods of the OTHER
-                            // privacy stored under the same name.
-                            let entry = class_def
-                                .methods
-                                .entry(resolved_method_name.clone())
-                                .or_default();
-                            entry.retain(|m| m.is_private != new_is_private);
-                            entry.push(def);
-                        }
-                    }
-                    // A method declared `is export` is recorded as an export of
-                    // the enclosing class so that `import ClassName` succeeds
-                    // (and exposes the method's sub-form name). This is mainly
-                    // used by operator methods such as `method infix:<as> is
-                    // export`, whose sub-form is importable.
-                    if *method_is_export && !self.suppress_exports {
-                        let tags = if method_export_tags.is_empty() {
-                            vec!["DEFAULT".to_string()]
-                        } else {
-                            method_export_tags.clone()
-                        };
-                        self.register_exported_var(
-                            name.to_string(),
-                            format!("&{}", resolved_method_name),
-                            tags,
-                        );
-                    }
-                    // Apply custom trait_mod:<is> for each non-builtin trait on methods
-                    if !method_custom_traits.is_empty() {
-                        let has_trait_mod = self.has_proto("trait_mod:<is>")
-                            || self.has_multi_candidates("trait_mod:<is>");
-                        if has_trait_mod {
-                            for (trait_name, trait_arg) in method_custom_traits {
-                                let mut trait_env = self.env.clone();
-                                // Add method lookup markers so .wrap stores in
-                                // method_wrap_chains (keyed by class+method).
-                                trait_env.insert(
-                                    "__mutsu_lookup_class".to_string(),
-                                    Value::str(name.to_string()),
-                                );
-                                trait_env.insert(
-                                    "__mutsu_lookup_method".to_string(),
-                                    Value::str(resolved_method_name.clone()),
-                                );
-                                trait_env.insert(
-                                    "__mutsu_lookup_candidate_idx".to_string(),
-                                    Value::Int(0),
-                                );
-                                let sub_val = Value::make_sub(
-                                    Symbol::intern(name),
-                                    Symbol::intern(&resolved_method_name),
-                                    effective_params.clone(),
-                                    effective_param_defs.clone(),
-                                    method_body.to_vec(),
-                                    *is_rw,
-                                    trait_env,
-                                );
-                                let trait_arg_val = if let Some(arg_expr) = trait_arg {
-                                    Some(self.eval_block_value(&[crate::ast::Stmt::Expr(
-                                        arg_expr.clone(),
-                                    )])?)
-                                } else {
-                                    None
-                                };
-                                let type_obj = self.resolve_type_object(trait_name);
-                                let mut args = vec![sub_val];
-                                if let Some(type_val) = type_obj {
-                                    args.push(type_val);
-                                    if let Some(arg_val) = trait_arg_val {
-                                        args.push(arg_val);
-                                    }
-                                    let _ = self.call_function("trait_mod:<is>", args);
-                                } else {
-                                    let named_val = if let Some(arg_val) = trait_arg_val {
-                                        Value::Pair(trait_name.clone(), Box::new(arg_val))
-                                    } else {
-                                        Value::Pair(trait_name.clone(), Box::new(Value::Bool(true)))
-                                    };
-                                    args.push(named_val);
-                                    let _ = self.call_function("trait_mod:<is>", args);
-                                }
-                            }
-                        }
-                    }
-                    // `handles` on a method: synthesize forwarder methods that
-                    // delegate to the return value of this method. E.g.
-                    //   method Str() handles 'uc' { 'x' }
-                    // registers a `uc` method that calls `self.Str.uc(|@_)`.
-                    if !method_handles.is_empty() {
-                        // Encode "method-based delegation" by prefixing the
-                        // source method name with `&`; the delegation dispatch
-                        // sites recognize this prefix and invoke the named
-                        // method on self to obtain the delegate.
-                        let source_attr_marker = format!("&{}", resolved_method_name);
-                        for spec in method_handles {
-                            match spec {
-                                HandleSpec::Name(target) => {
-                                    class_def
-                                        .methods
-                                        .entry(target.clone())
-                                        .or_default()
-                                        .push(make_delegation_method(&source_attr_marker, target));
-                                }
-                                HandleSpec::Rename { exposed, target } => {
-                                    class_def
-                                        .methods
-                                        .entry(exposed.clone())
-                                        .or_default()
-                                        .push(make_delegation_method(&source_attr_marker, target));
-                                }
-                                HandleSpec::Wildcard => {
-                                    class_def.wildcard_handles.push(source_attr_marker.clone());
-                                }
-                                HandleSpec::Regex(pattern) => {
-                                    class_def
-                                        .wildcard_handles
-                                        .push(format!("{}:regex:{}", source_attr_marker, pattern));
-                                }
-                                HandleSpec::Type(_) => {
-                                    // Method-based delegation via a type name
-                                    // is not yet supported; fall through.
-                                }
-                            }
-                        }
-                    }
-                    // `our method` also registers as a package-scoped sub
-                    if *is_our {
-                        let qualified_name = format!("{}::{}", name, resolved_method_name);
-                        let has_explicit_invocant = effective_param_defs
-                            .iter()
-                            .any(|p| p.is_invocant || p.traits.iter().any(|t| t == "invocant"));
-                        let (our_params, our_param_defs) = if has_explicit_invocant {
-                            (
-                                effective_params.clone(),
-                                effective_param_defs
-                                    .iter()
-                                    .filter(|p| p.name.as_str() != "self")
-                                    .cloned()
-                                    .collect(),
-                            )
-                        } else {
-                            // Prepend "self" as first param so the first argument
-                            // gets bound as `self` when calling this as a function.
-                            let mut our_params = vec!["self".to_string()];
-                            our_params.extend(
-                                effective_params
-                                    .iter()
-                                    .filter(|p| p.as_str() != "self")
-                                    .cloned(),
-                            );
-                            let self_param = crate::ast::ParamDef {
-                                name: "self".to_string(),
-                                default: None,
-                                multi_invocant: true,
-                                required: false,
-                                named: false,
-                                slurpy: false,
-                                double_slurpy: false,
-                                onearg: false,
-                                sigilless: false,
-                                type_constraint: None,
-                                literal_value: None,
-                                sub_signature: None,
-                                where_constraint: None,
-                                traits: Vec::new(),
-                                optional_marker: false,
-                                outer_sub_signature: None,
-                                code_signature: None,
-                                is_invocant: false,
-                                shape_constraints: None,
-                            };
-                            let mut our_param_defs = vec![self_param];
-                            our_param_defs.extend(
-                                effective_param_defs
-                                    .iter()
-                                    .filter(|p| {
-                                        !(p.is_invocant || p.traits.iter().any(|t| t == "invocant"))
-                                    })
-                                    .cloned(),
-                            );
-                            (our_params, our_param_defs)
-                        };
-                        let func_def = crate::ast::FunctionDef {
-                            package: Symbol::intern(name),
-                            name: Symbol::intern(&resolved_method_name),
-                            params: our_params,
-                            param_defs: our_param_defs,
-                            body: method_body.clone(),
-                            is_test_assertion: false,
-                            is_rw: *is_rw,
-                            is_raw: false,
-                            is_method: true,
-                            empty_sig: false,
-                            return_type: None,
-                            is_default: *is_default_candidate,
-                            deprecated_message: None,
-                        };
-                        self.registry_mut().functions.insert(
-                            Symbol::intern(&qualified_name),
-                            std::sync::Arc::new(func_def),
-                        );
-                    }
-                    // `my method` registers as a lexically-scoped function
-                    // (callable as `name(invocant)` inside the class body)
-                    if *is_my {
-                        let has_explicit_invocant = effective_param_defs
-                            .iter()
-                            .any(|p| p.is_invocant || p.traits.iter().any(|t| t == "invocant"));
-                        let (my_params, my_param_defs) = if has_explicit_invocant {
-                            (
-                                effective_params.clone(),
-                                effective_param_defs
-                                    .iter()
-                                    .filter(|p| p.name.as_str() != "self")
-                                    .cloned()
-                                    .collect(),
-                            )
-                        } else {
-                            let mut my_params = vec!["self".to_string()];
-                            my_params.extend(
-                                effective_params
-                                    .iter()
-                                    .filter(|p| p.as_str() != "self")
-                                    .cloned(),
-                            );
-                            let self_param = crate::ast::ParamDef {
-                                name: "self".to_string(),
-                                default: None,
-                                multi_invocant: true,
-                                required: false,
-                                named: false,
-                                slurpy: false,
-                                double_slurpy: false,
-                                onearg: false,
-                                sigilless: false,
-                                type_constraint: None,
-                                literal_value: None,
-                                sub_signature: None,
-                                where_constraint: None,
-                                traits: Vec::new(),
-                                optional_marker: false,
-                                outer_sub_signature: None,
-                                code_signature: None,
-                                is_invocant: false,
-                                shape_constraints: None,
-                            };
-                            let mut my_param_defs = vec![self_param];
-                            my_param_defs.extend(
-                                effective_param_defs
-                                    .iter()
-                                    .filter(|p| {
-                                        !(p.is_invocant || p.traits.iter().any(|t| t == "invocant"))
-                                    })
-                                    .cloned(),
-                            );
-                            (my_params, my_param_defs)
-                        };
-                        let func_def = crate::ast::FunctionDef {
-                            package: Symbol::intern(name),
-                            name: Symbol::intern(&resolved_method_name),
-                            params: my_params,
-                            param_defs: my_param_defs,
-                            body: method_body.clone(),
-                            is_test_assertion: false,
-                            is_rw: *is_rw,
-                            is_raw: false,
-                            is_method: true,
-                            empty_sig: false,
-                            return_type: None,
-                            is_default: *is_default_candidate,
-                            deprecated_message: None,
-                        };
-                        // Register under the short name (lexical scope)
-                        self.registry_mut().functions.insert(
-                            Symbol::intern(&resolved_method_name),
-                            std::sync::Arc::new(func_def.clone()),
-                        );
-                        // Also register under the qualified name for consistency
-                        let qualified_name = format!("{}::{}", name, resolved_method_name);
-                        self.registry_mut().functions.insert(
-                            Symbol::intern(&qualified_name),
-                            std::sync::Arc::new(func_def),
-                        );
-                        // Mark as my-scoped so it doesn't appear in the package stash
-                        self.mark_my_scoped_package_item(qualified_name);
-                    }
-                }
-                Stmt::DoesDecl { name: role_name } => {
-                    let role_name_str = role_name.resolve();
-                    if !self.registry().roles.contains_key(&role_name_str)
-                        && matches!(
-                            role_name_str.as_str(),
-                            "Real"
-                                | "Numeric"
-                                | "Cool"
-                                | "Any"
-                                | "Mu"
-                                | "Positional"
-                                | "Associative"
-                        )
-                    {
-                        if !class_def.parents.iter().any(|p| p == &role_name_str) {
-                            class_def.parents.insert(0, role_name_str.clone());
-                            class_def.mro.clear();
-                        }
-                        continue;
-                    }
-                    let role = self
-                        .registry()
-                        .roles
-                        .get(&role_name_str)
-                        .cloned()
-                        .ok_or_else(|| {
-                            RuntimeError::new(format!("Unknown role: {}", role_name_str))
-                        })?;
-                    if role.is_stub_role {
-                        return Err(RuntimeError::typed_msg(
-                            "X::Role::Parametric::NoSuchCandidate",
-                            "No matching candidate found for the parametric role",
-                        ));
-                    }
-                    // Look up the role's language revision for submethod composition rules.
-                    let role_lang_rev_does = self
-                        .type_metadata
-                        .get(&role_name_str)
-                        .and_then(|m| m.get("language-revision"))
-                        .map(|v| v.to_string_value())
-                        .unwrap_or_else(|| "c".to_string());
-                    let compose_submethods_does =
-                        class_lang_rev == "c" && role_lang_rev_does == "c";
-                    for attr in &role.attributes {
-                        if !class_def.attributes.iter().any(|(n, ..)| n == &attr.0) {
-                            class_def.attributes.push(attr.clone());
-                        }
-                    }
-                    for (mname, overloads) in role.methods {
-                        let composed: Vec<MethodDef> = overloads
-                            .into_iter()
-                            .filter(|md| !md.is_my || (md.is_submethod && compose_submethods_does))
-                            .map(|mut md| {
-                                if md.original_role.is_none() {
-                                    md.original_role = md.role_origin.clone();
-                                }
-                                md.role_origin = Some(role_name_str.clone());
-                                md
-                            })
-                            .collect();
-                        if composed.is_empty() {
-                            continue;
-                        }
-                        class_def.methods.entry(mname).or_default().extend(composed);
-                    }
-                    // Transfer wildcard handles from role to class
-                    for wh in &role.wildcard_handles {
-                        if !class_def.wildcard_handles.contains(wh) {
-                            class_def.wildcard_handles.push(wh.clone());
-                        }
-                    }
-                    if !class_def.parents.iter().any(|p| p == &role_name_str) {
-                        // Keep role composition visible in MRO introspection.
-                        class_def.parents.insert(0, role_name_str.clone());
-                        class_def.mro.clear();
-                    }
-                    // Transfer role's own parents (from `is` declarations) to the class
-                    if let Some(rparents) =
-                        self.registry().role_parents.get(&role_name_str).cloned()
-                    {
-                        for rp in rparents {
-                            let rp_base = rp.split_once('[').map(|(b, _)| b).unwrap_or(rp.as_str());
-                            if self.registry().classes.contains_key(rp_base)
-                                && !class_def.parents.iter().any(|p| p == &rp)
-                            {
-                                class_def.parents.push(rp.clone());
-                                class_def.mro.clear();
-                            }
-                        }
-                    }
-                }
-                Stmt::TrustsDecl {
-                    name: trusted_class,
-                } => {
-                    self.registry_mut()
-                        .class_trusts
-                        .entry(name.to_string())
-                        .or_default()
-                        .insert(trusted_class.resolve());
-                }
-                // our &baz ::= &bar  — alias a method under a new name
-                Stmt::VarDecl {
-                    name: var_name,
-                    expr: Expr::CodeVar(source_name),
-                    ..
-                } if var_name.starts_with('&') => {
-                    let alias = var_name.trim_start_matches('&').to_string();
-                    if let Some(overloads) = class_def.methods.get(source_name).cloned() {
-                        class_def.methods.insert(alias, overloads);
-                    }
-                    // Also execute the statement so the code variable is set
-                    self.registry_mut()
-                        .classes
-                        .insert(name.to_string(), class_def.clone());
-                    self.run_block_raw(std::slice::from_ref(stmt))?;
-                    for outer_name in saved_env.keys() {
-                        let class_scoped_name = format!("{}::{}", name, outer_name);
-                        if let Some(updated) = self.env.get(&class_scoped_name).cloned() {
-                            self.env.insert_sym(*outer_name, updated);
-                        }
-                    }
-                    if let Some(updated) = self.registry().classes.get(name).cloned() {
-                        class_def = updated;
-                    }
-                }
-                Stmt::ProtoDecl {
-                    name: proto_name,
-                    param_defs,
-                    body: proto_body,
-                    is_method: true,
-                    ..
-                } => {
-                    let method_name = proto_name.resolve();
-                    let effective_param_defs = Self::effective_method_param_defs(param_defs, false);
-                    let effective_params: Vec<String> = effective_param_defs
-                        .iter()
-                        .map(|p| p.name.clone())
-                        .collect();
-                    let fdef = FunctionDef {
-                        package: Symbol::intern(name),
-                        name: *proto_name,
-                        params: effective_params,
-                        param_defs: effective_param_defs,
-                        body: proto_body.clone(),
-                        is_test_assertion: false,
-                        is_rw: false,
-                        is_raw: false,
-                        is_method: true,
-                        empty_sig: false,
-                        return_type: None,
-                        is_default: false,
-                        deprecated_message: None,
-                    };
-                    self.registry_mut()
-                        .proto_methods
-                        .insert((name.to_string(), method_name), fdef);
-                }
-                _ => {
-                    // BEGIN phasers and EVAL calls in class bodies may fail
-                    // (e.g. `BEGIN EVAL q[has $.x]` or `EVAL q[has $.x]`).
-                    // Swallow errors from these so the class still registers.
-                    let is_swallowable = matches!(
-                        stmt,
-                        Stmt::Phaser {
-                            kind: PhaserKind::Begin,
-                            ..
-                        }
-                    ) || matches!(
-                        stmt,
-                        Stmt::Call { name: fn_name, .. }
-                            if fn_name.resolve() == "EVAL"
-                    ) || matches!(
-                        stmt,
-                        Stmt::Expr(Expr::Call { name: fn_name, .. })
-                            if fn_name.resolve() == "EVAL"
-                    );
-                    self.registry_mut()
-                        .classes
-                        .insert(name.to_string(), class_def.clone());
-                    let result = self.run_block_raw(std::slice::from_ref(stmt));
-                    if let Err(e) = result {
-                        if !is_swallowable {
-                            return Err(e);
-                        }
-                    } else {
-                        for outer_name in saved_env.keys() {
-                            let class_scoped_name = format!("{}::{}", name, outer_name);
-                            if let Some(updated) = self.env.get(&class_scoped_name).cloned() {
-                                self.env.insert_sym(*outer_name, updated);
-                            }
-                        }
-                    }
-                    if let Some(updated) = self.registry().classes.get(name).cloned() {
-                        class_def = updated;
-                    }
-                }
-            }
-            // Check if any new functions were registered under the class package
-            // during body processing (e.g., class-scoped subs).
-            // Only count functions that are actual subs (not methods, which are
-            // registered via MethodDecl and stored in the class methods table).
-            if let Stmt::SubDecl { name: sub_name, .. } = stmt {
-                let fq = format!("{}::{}", name, sub_name);
-                if self.registry().functions.contains_key(&Symbol::intern(&fq))
-                    && !saved_functions_keys.contains(&fq)
-                {
-                    self.registry_mut()
-                        .class_subs
-                        .entry(name.to_string())
-                        .or_default()
-                        .insert(fq, Value::Bool(true));
-                }
-            }
-            self.registry_mut()
-                .classes
-                .insert(name.to_string(), class_def.clone());
-        }
-        // Fire class-body LEAVE phasers in LIFO order now that the body scope
-        // is being left. They run while the class package/env are still active
-        // so their bodies can see body-scoped variables; writes to outer
-        // variables persist because the class-body env is not rolled back on
-        // the success path.
-        for body in class_leave_phasers.iter().rev() {
-            self.run_block_raw(body)?;
-            for outer_name in saved_env.keys() {
-                let class_scoped_name = format!("{}::{}", name, outer_name);
-                if let Some(updated) = self.env.get(&class_scoped_name).cloned() {
-                    self.env.insert_sym(*outer_name, updated);
-                }
-            }
-        }
-        self.set_current_package(saved_package);
-        if let Err(err) = self.resolve_class_stub_requirements(name, &mut class_def) {
-            restore_previous_state(self);
-            return Err(err);
-        }
-        if let Err(err) = self.detect_unresolved_role_method_conflicts(name, &class_def) {
-            restore_previous_state(self);
-            return Err(err);
-        }
-        self.registry_mut()
-            .classes
-            .insert(name.to_string(), class_def);
-        let mut stack = Vec::new();
-        if let Err(err) = self.compute_class_mro(name, &mut stack) {
-            restore_previous_state(self);
-            return Err(err);
-        }
-        // Validate that all self!method() calls reference existing private methods
-        if let Err(err) = self.validate_private_method_existence(name) {
-            restore_previous_state(self);
-            return Err(err);
-        }
-        Ok(deferred_custom_traits)
-    }
-
-    /// Build the error for `augment role RoleName { ... }`. Roles are always
-    /// closed, so augmenting one is illegal — but if the role does not exist at
-    /// all, Raku reports that first (X::Augment::NoSuchType), matching the
-    /// ordering used for `augment class`.
-    pub(crate) fn augment_role_error(&self, name: &str) -> RuntimeError {
-        // Built-in roles that are closed but not present in the user registry.
-        const BUILTIN_ROLES: &[&str] = &[
-            "Positional",
-            "Associative",
-            "Callable",
-            "Iterable",
-            "Numeric",
-            "Real",
-            "Stringy",
-            "Mixy",
-            "Setty",
-            "Baggy",
-            "Blob",
-            "Buf",
-        ];
-        let exists = self.registry().roles.contains_key(name)
-            || self.registry().classes.contains_key(name)
-            || crate::runtime::utils::is_known_type_constraint(name)
-            || BUILTIN_ROLES.contains(&name);
-        if exists {
-            let message = format!("Cannot augment {} because it is closed", name);
-            let mut attrs = std::collections::HashMap::new();
-            attrs.insert("message".to_string(), Value::str(message.clone()));
-            let ex = Value::make_instance(
-                crate::symbol::Symbol::intern("X::Syntax::Augment::Illegal"),
-                attrs,
-            );
-            let mut err = RuntimeError::new(message);
-            err.exception = Some(Box::new(ex));
-            err
-        } else {
-            let message = format!("You tried to augment role {}, but it does not exist", name);
-            let mut attrs = std::collections::HashMap::new();
-            attrs.insert("message".to_string(), Value::str(message.clone()));
-            attrs.insert("package-kind".to_string(), Value::str("role".to_string()));
-            attrs.insert("package".to_string(), Value::str(name.to_string()));
-            let ex = Value::make_instance(
-                crate::symbol::Symbol::intern("X::Augment::NoSuchType"),
-                attrs,
-            );
-            let mut err = RuntimeError::new(message);
-            err.exception = Some(Box::new(ex));
-            err
-        }
-    }
-
-    /// Augment an existing class by adding methods (and attributes) from the body.
-    /// This implements `augment class ClassName { ... }` (monkey-patching).
-    pub(crate) fn augment_class(&mut self, name: &str, body: &[Stmt]) -> Result<(), RuntimeError> {
-        self.clear_private_zeroarg_method_cache();
-        // Check if the class exists (user-defined or builtin)
-        let is_builtin = !self.registry().classes.contains_key(name);
-        if is_builtin {
-            // For builtin types, create a minimal class def so we can add methods
-            const BUILTIN_TYPES: &[&str] = &[
-                "Mu",
-                "Any",
-                "Cool",
-                "Int",
-                "Num",
-                "Str",
-                "Bool",
-                "Rat",
-                "FatRat",
-                "Complex",
-                "Array",
-                "Hash",
-                "List",
-                "Map",
-                "Set",
-                "Bag",
-                "Mix",
-                "Range",
-                "Pair",
-                "IO",
-                "IO::Path",
-                "IO::Handle",
-                "Regex",
-                "Match",
-                "Junction",
-                "Exception",
-                "Failure",
-                "Version",
-                "Nil",
-                "Block",
-                "Code",
-                "Routine",
-                "Sub",
-                "Method",
-                "Seq",
-                "Slip",
-                "Whatever",
-                "WhateverCode",
-                "HyperWhatever",
-                "Callable",
-                "Numeric",
-                "Real",
-                "Stringy",
-                "Positional",
-                "Associative",
-                "Order",
-                "Endian",
-                "Proc",
-                "Capture",
-            ];
-            if !BUILTIN_TYPES.contains(&name) {
-                let message = format!("You tried to augment class {}, but it does not exist", name);
-                let mut attrs = std::collections::HashMap::new();
-                attrs.insert("message".to_string(), Value::str(message.clone()));
-                attrs.insert("package-kind".to_string(), Value::str("class".to_string()));
-                attrs.insert("package".to_string(), Value::str(name.to_string()));
-                let ex = Value::make_instance(
-                    crate::symbol::Symbol::intern("X::Augment::NoSuchType"),
-                    attrs,
-                );
-                let mut err = RuntimeError::new(message);
-                err.exception = Some(Box::new(ex));
-                return Err(err);
-            }
-            // Create a minimal entry for the builtin type
-            self.registry_mut()
-                .classes
-                .entry(name.to_string())
-                .or_default();
-        }
-
-        let saved_package = self.current_package();
-        self.set_current_package(name.to_string());
-
-        // Process body statements and add methods/attributes to the existing class
-        let flattened_body: Vec<&Stmt> = body
-            .iter()
-            .flat_map(|s| match s {
-                Stmt::SyntheticBlock(inner) => inner.iter().collect::<Vec<_>>(),
-                other => vec![other],
-            })
-            .collect();
-        for stmt in flattened_body {
-            match stmt {
-                Stmt::MethodDecl {
-                    name: method_name,
-                    name_expr,
-                    param_defs,
-                    body: method_body,
-                    multi,
-                    is_rw,
-                    is_private,
-                    is_my,
-                    is_submethod,
-                    return_type,
-                    is_default_candidate,
-                    ..
-                } => {
-                    let resolved_method_name = if let Some(expr) = name_expr {
-                        self.eval_block_value(&[Stmt::Expr(expr.clone())])?
-                            .to_string_value()
-                    } else {
-                        method_name.resolve()
-                    };
-                    let mut effective_param_defs =
-                        Self::effective_method_param_defs(param_defs, false);
-                    // Auto-detect @_ usage in methods without explicit signatures
-                    if param_defs.is_empty() {
-                        let (use_positional, _) = Self::auto_signature_uses(method_body);
-                        if use_positional && !effective_param_defs.iter().any(|pd| pd.name == "@_")
-                        {
-                            let insert_pos = effective_param_defs
-                                .iter()
-                                .position(|pd| pd.name.starts_with('%') && pd.slurpy)
-                                .unwrap_or(effective_param_defs.len());
-                            effective_param_defs.insert(
-                                insert_pos,
-                                ParamDef {
-                                    name: "@_".to_string(),
-                                    default: None,
-                                    multi_invocant: true,
-                                    required: false,
-                                    named: false,
-                                    slurpy: true,
-                                    double_slurpy: false,
-                                    onearg: false,
-                                    sigilless: false,
-                                    type_constraint: None,
-                                    literal_value: None,
-                                    sub_signature: None,
-                                    where_constraint: None,
-                                    traits: Vec::new(),
-                                    optional_marker: false,
-                                    outer_sub_signature: None,
-                                    code_signature: None,
-                                    is_invocant: false,
-                                    shape_constraints: None,
-                                },
-                            );
-                        }
-                    }
-                    let effective_params: Vec<String> = effective_param_defs
-                        .iter()
-                        .map(|p| p.name.clone())
-                        .collect();
-                    let def = MethodDef {
-                        params: effective_params,
-                        param_defs: effective_param_defs,
-                        body: std::sync::Arc::new(method_body.clone()),
-                        is_rw: *is_rw,
-                        is_private: *is_private,
-                        is_multi: *multi,
-                        is_my: *is_my,
-                        role_origin: None,
-                        original_role: None,
-                        return_type: return_type.clone(),
-                        compiled_code: None,
-                        delegation: None,
-                        is_default: *is_default_candidate,
-                        deprecated_message: None,
-                        is_submethod: *is_submethod,
-                    };
-                    if let Some(class_def) = self.registry_mut().classes.get_mut(name) {
-                        if *multi {
-                            class_def
-                                .methods
-                                .entry(resolved_method_name)
-                                .or_default()
-                                .push(def);
-                        } else {
-                            // Check for duplicate non-multi method definition.
-                            // Only error if the existing method was defined in
-                            // this class (not composed from a role).
-                            if let Some(existing) = class_def.methods.get(&resolved_method_name) {
-                                let all_from_role =
-                                    existing.iter().all(|m| m.role_origin.is_some());
-                                if !all_from_role {
-                                    return Err(RuntimeError::new(format!(
-                                        "Package '{}' already has a method '{}' (did you mean to declare a multi method?)",
-                                        name, resolved_method_name
-                                    )));
-                                }
-                            }
-                            class_def.methods.insert(resolved_method_name, vec![def]);
-                        }
-                    }
-                }
-                Stmt::HasDecl {
-                    name: attr_name,
-                    is_public,
-                    default,
-                    is_rw,
-                    is_readonly: _,
-                    type_constraint,
-                    type_smiley: _,
-                    is_required,
-                    sigil,
-                    handles,
-                    where_constraint,
-                    is_alias,
-                    is_our: _,
-                    is_my: _,
-                    is_default: _,
-                    is_type: _,
-                    deprecated_message: _,
-                    is_built: _,
-                    unknown_traits: _,
-                } => {
-                    let attr_name_str = attr_name.resolve();
-                    let attr_var_name = if *is_public {
-                        format!(".{}", attr_name_str)
-                    } else {
-                        format!("!{}", attr_name_str)
-                    };
-                    // Resolve handles before taking mutable borrow on class_def
-                    let resolved = self.resolve_handle_specs_to_names(handles, &attr_var_name);
-                    if let Some(class_def) = self.registry_mut().classes.get_mut(name) {
-                        class_def.attributes.push((
-                            attr_name_str.clone(),
-                            *is_public,
-                            default.clone(),
-                            *is_rw,
-                            is_required.clone(),
-                            *sigil,
-                            where_constraint.as_ref().map(|wc| wc.as_ref().clone()),
-                        ));
-                        if *is_alias {
-                            class_def.alias_attributes.insert(attr_name_str.clone());
-                        }
-                        if let Some(tc) = type_constraint {
-                            class_def
-                                .attribute_types
-                                .insert(attr_name_str.clone(), tc.clone());
-                        }
-                        apply_resolved_handles(
-                            &resolved,
-                            &mut class_def.methods,
-                            &mut class_def.wildcard_handles,
-                        );
-                    }
-                }
-                // Sub declarations in the class body should persist across scope
-                // boundaries so that methods can call them. Use run_block_raw
-                // instead of eval_block_value to avoid function registry rollback.
-                Stmt::SubDecl { .. } => {
-                    let _ = self.run_block_raw(std::slice::from_ref(stmt));
-                }
-                other => {
-                    let _ = self.eval_block_value(std::slice::from_ref(other));
-                }
-            }
-        }
-
-        self.set_current_package(saved_package);
-        Ok(())
-    }
-
-    pub(crate) fn register_role_decl(
-        &mut self,
-        name: &str,
-        type_params: &[String],
-        type_param_defs: &[ParamDef],
-        body: &[Stmt],
-        role_is_rw: bool,
-    ) -> Result<(), RuntimeError> {
-        self.clear_private_zeroarg_method_cache();
-
-        // Check for our-scoped declarations inside the role body.
-        // In Raku, class/subset/enum/constant/role are implicitly our-scoped,
-        // and explicit `our sub/method/variable` are also forbidden inside roles.
-        let check_body: Vec<&Stmt> = body
-            .iter()
-            .flat_map(|s| match s {
-                Stmt::SyntheticBlock(inner) => inner.iter().collect::<Vec<_>>(),
-                other => vec![other],
-            })
-            .collect();
-        for stmt in &check_body {
-            let declaration = match stmt {
-                // A `my class` inside a role is a lexically-scoped class private to the
-                // role and is allowed; only an implicitly our-scoped `class` is forbidden.
-                Stmt::ClassDecl {
-                    is_lexical: false, ..
-                } => Some("class"),
-                Stmt::ClassDecl { .. } => None,
-                Stmt::SubsetDecl { .. } => Some("subset"),
-                Stmt::EnumDecl { .. } => Some("enum"),
-                // A `my role` is lexically scoped and private to the role body, which
-                // is allowed (like `my class`); only an implicitly our-scoped `role` is
-                // forbidden.
-                Stmt::RoleDecl { custom_traits, .. }
-                    if custom_traits.iter().any(|(t, _)| t == "__my_scoped") =>
-                {
-                    None
-                }
-                Stmt::RoleDecl { .. } => Some("role"),
-                Stmt::VarDecl {
-                    is_our: true,
-                    custom_traits,
-                    ..
-                } => {
-                    if custom_traits.iter().any(|(t, _)| t == "__constant") {
-                        Some("constant")
-                    } else {
-                        Some("variable")
-                    }
-                }
-                Stmt::SubDecl { custom_traits, .. }
-                    if custom_traits.iter().any(|(t, _)| t == "__our_scoped") =>
-                {
-                    Some("sub")
-                }
-                Stmt::MethodDecl { is_our: true, .. } => Some("method"),
-                _ => None,
-            };
-            if let Some(decl) = declaration {
-                let mut attrs = std::collections::HashMap::new();
-                attrs.insert("declaration".to_string(), Value::str(decl.to_string()));
-                attrs.insert(
-                    "message".to_string(),
-                    Value::str(format!(
-                        "Cannot declare our-scoped {} inside of a role",
-                        decl
-                    )),
-                );
-                return Err(RuntimeError::typed("X::Declaration::OurScopeInRole", attrs));
-            }
-        }
-
-        for param_def in type_param_defs {
-            if param_def.name == "__type_only__"
-                && let Some(type_name) = param_def.type_constraint.as_deref()
-                && !type_name.starts_with("::")
-                && !self.is_resolvable_type(type_name)
-            {
-                let mut attrs = std::collections::HashMap::new();
-                attrs.insert("type".to_string(), Value::str(type_name.to_string()));
-                attrs.insert(
-                    "message".to_string(),
-                    Value::str(format!(
-                        "Invalid type '{}' used in role parameter list",
-                        type_name
-                    )),
-                );
-                return Err(RuntimeError::typed("X::Parameter::InvalidType", attrs));
-            }
-        }
-
-        // If this is a stub declaration (body is `...`, `!!!`, or `???`)
-        // and a real (non-stub) role already exists under this name, treat
-        // the stub as a forward declaration / no-op — don't register a new
-        // stub candidate that would shadow the real one.
-        let is_stub_body = body.iter().any(|s| {
-            matches!(s, Stmt::Expr(Expr::Call { name, .. })
-                if name == "__mutsu_stub_die" || name == "__mutsu_stub_warn")
-        });
-        if is_stub_body
-            && type_params.is_empty()
-            && self
-                .registry()
-                .roles
-                .get(name)
-                .is_some_and(|existing| !existing.is_stub_role)
-        {
-            return Ok(());
-        }
-        // Clean up stale punned class entry for this role name.
-        self.registry_mut().classes.remove(name);
-        self.registry_mut().hidden_classes.remove(name);
-        self.registry_mut().class_composed_roles.remove(name);
-        // When registering a parametric variant of an existing non-parametric role
-        // (forming a role group), save the non-parametric role's parents so we can
-        // restore them after the parametric variant adds its own parents.
-        let prev_parents = if !type_params.is_empty()
-            && self
-                .registry()
-                .roles
-                .get(name)
-                .is_some_and(|existing| !existing.is_stub_role)
-        {
-            self.registry().role_parents.get(name).cloned()
-        } else {
-            None
-        };
-        self.registry_mut().role_parents.remove(name);
-        self.registry_mut().role_hides.remove(name);
-        let mut role_def = RoleDef {
-            attributes: Vec::new(),
-            methods: HashMap::new(),
-            // A yada-body forward declaration (`role Foo { ... }`) is a stub role:
-            // mark it so the real definition that follows replaces it instead of
-            // being treated as a second, conflicting role of the same name (which
-            // made transitively-composed methods like `add_route` look like a
-            // cross-role X::Role::Composition::Conflict).
-            is_stub_role: is_stub_body,
-            is_hidden: false,
-            is_rw: role_is_rw,
-            captured_env: None,
-            wildcard_handles: Vec::new(),
-            role_id: super::next_role_id(),
-            attribute_conflicts: Vec::new(),
-            own_attribute_names: HashSet::new(),
-            deferred_body_stmts: Vec::new(),
-            deferred_custom_traits: Vec::new(),
-        };
-        let is_parametric = !type_params.is_empty();
-        let flattened_body: Vec<&Stmt> = body
-            .iter()
-            .flat_map(|s| match s {
-                Stmt::SyntheticBlock(inner) => inner.iter().collect::<Vec<_>>(),
-                other => vec![other],
-            })
-            .collect();
-        // Pre-scan: collect attribute names declared in this role body.
-        let mut role_own_attrs: HashSet<String> = HashSet::new();
-        for stmt in &flattened_body {
-            if let Stmt::HasDecl {
-                name: attr_name, ..
-            } = stmt
-            {
-                role_own_attrs.insert(attr_name.resolve());
-            }
-        }
-        let role_attr_ctx = AttrValidationCtx {
-            attrs: &role_own_attrs,
-            pkg_name: name,
-            pkg_kind: "role",
-        };
-        for stmt in flattened_body {
-            match stmt {
-                Stmt::HasDecl {
-                    name: attr_name,
-                    is_public,
-                    default,
-                    handles,
-                    is_rw,
-                    is_readonly,
-                    type_constraint: _,
-                    type_smiley: _,
-                    is_required,
-                    sigil,
-                    where_constraint,
-                    is_alias: _,
-                    is_our: _,
-                    is_my: _,
-                    is_default: _,
-                    is_type: _,
-                    deprecated_message: _,
-                    is_built: _,
-                    unknown_traits: _,
-                } => {
-                    let attr_name_str = attr_name.resolve();
-                    role_def.own_attribute_names.insert(attr_name_str.clone());
-                    // Check if this attribute already exists from a composed role
-                    if let Some(existing) = role_def
-                        .attributes
-                        .iter()
-                        .find(|(n, ..)| n == &attr_name_str)
-                    {
-                        // The attribute already exists from a parent role composition.
-                        // Record the conflict; the existing one came from a composed role.
-                        // We need to figure out which role contributed it.
-                        let parent_role = self
-                            .registry()
-                            .role_parents
-                            .get(name)
-                            .and_then(|parents| {
-                                parents.iter().find(|p| {
-                                    let base =
-                                        p.split_once('[').map(|(b, _)| b).unwrap_or(p.as_str());
-                                    self.registry().roles.get(base).is_some_and(|r| {
-                                        r.attributes.iter().any(|(n, ..)| n == &attr_name_str)
-                                    })
-                                })
-                            })
-                            .cloned()
-                            .unwrap_or_else(|| "unknown".to_string());
-                        let _ = existing;
-                        role_def.attribute_conflicts.push((
-                            attr_name_str.clone(),
-                            name.to_string(),
-                            parent_role,
-                        ));
-                    }
-                    // Apply role-level `is rw`: same logic as class_is_rw
-                    // `is readonly` on individual attributes overrides `is rw` on the role
-                    let effective_is_rw = !*is_readonly && (*is_rw || (role_is_rw && *is_public));
-                    role_def.attributes.push((
-                        attr_name_str.clone(),
-                        *is_public,
-                        default.clone(),
-                        effective_is_rw,
-                        is_required.clone(),
-                        *sigil,
-                        where_constraint.as_ref().map(|wc| wc.as_ref().clone()),
-                    ));
-                    let attr_var_name = if *is_public {
-                        format!(".{}", attr_name_str)
-                    } else {
-                        format!("!{}", attr_name_str)
-                    };
-                    self.apply_handle_specs_to_role(handles, &attr_var_name, &mut role_def);
-                }
-                Stmt::DoesDecl { name: role_name } => {
-                    if *role_name == "__mutsu_role_hidden__" {
-                        role_def.is_hidden = true;
-                        continue;
-                    }
-                    let role_name_str = role_name.resolve();
-                    if let Some(hidden_name) = role_name_str.strip_prefix("__mutsu_role_hides__") {
-                        // Track hidden class relationship for this role
-                        self.registry_mut()
-                            .role_hides
-                            .entry(name.to_string())
-                            .or_default()
-                            .push(hidden_name.to_string());
-                        continue;
-                    }
-                    if self.registry().classes.contains_key(&role_name_str) {
-                        self.registry_mut()
-                            .role_parents
-                            .entry(name.to_string())
-                            .or_default()
-                            .push(role_name_str);
-                        continue;
-                    }
-                    let base_role_name = role_name_str
-                        .split_once('[')
-                        .map(|(b, _)| b)
-                        .unwrap_or(role_name_str.as_str());
-                    if type_params.iter().any(|tp| tp == base_role_name)
-                        || (!self.registry().roles.contains_key(base_role_name)
-                            && matches!(
-                                base_role_name,
-                                "Real"
-                                    | "Numeric"
-                                    | "Cool"
-                                    | "Any"
-                                    | "Mu"
-                                    | "Positional"
-                                    | "Associative"
-                            ))
-                    {
-                        self.registry_mut()
-                            .role_parents
-                            .entry(name.to_string())
-                            .or_default()
-                            .push(role_name_str);
-                        continue;
-                    }
-                    let role = match self.registry().roles.get(base_role_name).cloned() {
-                        Some(r) => r,
-                        None => {
-                            // If trait_mod:<is> is defined and this is a lowercase name,
-                            // defer to custom trait dispatch instead of erroring.
-                            if (self.has_proto("trait_mod:<is>")
-                                || self.has_multi_candidates("trait_mod:<is>"))
-                                && role_name_str
-                                    .chars()
-                                    .next()
-                                    .is_some_and(|c| c.is_ascii_lowercase())
-                            {
-                                role_def
-                                    .deferred_custom_traits
-                                    .push(role_name_str.to_string());
-                                continue;
-                            }
-                            return Err(RuntimeError::new(format!(
-                                "Unknown role: {}",
-                                role_name_str
-                            )));
-                        }
-                    };
-                    if role.is_stub_role {
-                        return Err(RuntimeError::typed_msg(
-                            "X::Role::Parametric::NoSuchCandidate",
-                            "No matching candidate found for the parametric role",
-                        ));
-                    }
-                    self.registry_mut()
-                        .role_parents
-                        .entry(name.to_string())
-                        .or_default()
-                        .push(role_name_str.clone());
-                    // Use resolve_role_candidate to properly handle named
-                    // arguments and default values in parameterized role
-                    // composition.
-                    let type_subs: Vec<(String, String)> =
-                        if let Some((_, resolved_param_names, resolved_values)) =
-                            self.resolve_role_candidate(&role_name_str)?
-                        {
-                            // Store the resolved param bindings so they are
-                            // available when the child role is punned to a class
-                            // and methods referencing role params are dispatched.
-                            {
-                                let mut registry = self.registry_mut();
-                                let bindings = registry
-                                    .class_role_param_bindings
-                                    .entry(name.to_string())
-                                    .or_default();
-                                for (p, v) in
-                                    resolved_param_names.iter().zip(resolved_values.iter())
-                                {
-                                    bindings.insert(p.clone(), v.clone());
-                                }
-                            }
-                            resolved_param_names
-                                .iter()
-                                .zip(resolved_values.iter())
-                                .map(|(p, v)| (p.clone(), type_value_name(v)))
-                                .collect()
-                        } else if let Some(parent_type_params) =
-                            self.registry().role_type_params.get(base_role_name)
-                        {
-                            if let Some(bracket_start) = role_name_str.find('[') {
-                                let args_str =
-                                    &role_name_str[bracket_start + 1..role_name_str.len() - 1];
-                                let type_args = parse_role_type_args(args_str);
-                                parent_type_params
-                                    .iter()
-                                    .zip(type_args.iter())
-                                    .map(|(p, a)| (p.clone(), a.clone()))
-                                    .collect()
-                            } else {
-                                Vec::new()
-                            }
-                        } else {
-                            Vec::new()
-                        };
-                    for attr in &role.attributes {
-                        if role_def.attributes.iter().any(|(n, ..)| n == &attr.0) {
-                            // Already present. Only a real conflict if both
-                            // sides declared it directly (vs. inherited from
-                            // a shared ancestor in a diamond). Skip otherwise.
-                            let parent_owns = role.own_attribute_names.contains(&attr.0);
-                            let current_owns = role_def.own_attribute_names.contains(&attr.0);
-                            if parent_owns && current_owns {
-                                role_def.attribute_conflicts.push((
-                                    attr.0.clone(),
-                                    name.to_string(),
-                                    base_role_name.to_string(),
-                                ));
-                            }
-                        } else {
-                            role_def.attributes.push(attr.clone());
-                        }
-                    }
-                    for (mname, overloads) in role.methods {
-                        // Skip methods declared with `my` scope -- role-private
-                        // Submethods (is_submethod=true) ARE composed even though
-                        // they have is_my=true.
-                        let non_my_overloads: Vec<MethodDef> = overloads
-                            .into_iter()
-                            .filter(|md| !md.is_my || md.is_submethod)
-                            .collect();
-                        if non_my_overloads.is_empty() {
-                            continue;
-                        }
-                        let composed: Vec<MethodDef> = if type_subs.is_empty() {
-                            non_my_overloads
-                                .into_iter()
-                                .map(|mut md| {
-                                    if md.original_role.is_none() {
-                                        md.original_role = md.role_origin.clone();
-                                    }
-                                    md.role_origin = Some(base_role_name.to_string());
-                                    md
-                                })
-                                .collect()
-                        } else {
-                            non_my_overloads
-                                .iter()
-                                .map(|md| {
-                                    let mut method =
-                                        substitute_type_params_in_method(md, &type_subs);
-                                    if method.original_role.is_none() {
-                                        method.original_role = method.role_origin.clone();
-                                    }
-                                    method.role_origin = Some(base_role_name.to_string());
-                                    method
-                                })
-                                .collect()
-                        };
-                        role_def.methods.entry(mname).or_default().extend(composed);
-                    }
-                }
-                Stmt::MethodDecl {
-                    name: method_name,
-                    name_expr,
-                    params,
-                    param_defs,
-                    body: method_body,
-                    multi,
-                    is_rw,
-                    is_private,
-                    is_our: _,
-                    is_my,
-                    is_submethod,
-                    our_variable_form: _,
-                    return_type,
-                    is_default_candidate,
-                    deprecated_message: _,
-                    handles: method_handles,
-                    custom_traits: _,
-                    is_export: _,
-                    export_tags: _,
-                } => {
-                    // Validate that $!attr references in the method body are declared
-                    // in this role (same check as for class methods).
-                    Self::validate_attr_declared_in_class(&role_attr_ctx, method_body)?;
-                    // Validate that type constraints in method parameters are resolvable.
-                    // Undeclared types like A::C should throw X::Parameter::InvalidType.
-                    for pd in param_defs {
-                        if let Some(tc) = pd.type_constraint.as_deref() {
-                            // Skip type captures (::T), invocant markers, and role type params
-                            if tc.starts_with("::")
-                                || tc == "__invocant__"
-                                || type_params.iter().any(|tp| tp == tc)
-                            {
-                                continue;
-                            }
-                            if !self.is_resolvable_type(tc) {
-                                let mut attrs = std::collections::HashMap::new();
-                                attrs.insert("type".to_string(), Value::str(tc.to_string()));
-                                attrs.insert(
-                                    "message".to_string(),
-                                    Value::str(format!(
-                                        "Invalid typename '{}' in parameter declaration.",
-                                        tc
-                                    )),
-                                );
-                                return Err(RuntimeError::typed(
-                                    "X::Parameter::InvalidType",
-                                    attrs,
-                                ));
-                            }
-                        }
-                    }
-                    // Stub multi methods (body is `{...}`) that use ::?CLASS
-                    // must be implemented by the composing class.
-                    // Non-stub multi methods with ::?CLASS are fine.
-                    let body_is_stub = {
-                        let filtered: Vec<_> = method_body
-                            .iter()
-                            .filter(|s| !matches!(s, Stmt::SetLine(_)))
-                            .collect();
-                        filtered.len() == 1
-                            && matches!(
-                                filtered[0],
-                                Stmt::Expr(Expr::Call { name, .. })
-                                    if name == "__mutsu_stub_die"
-                                        || name == "__mutsu_stub_warn"
-                            )
-                    };
-                    if *multi
-                        && body_is_stub
-                        && (param_defs.iter().any(|pd| {
-                            pd.type_constraint
-                                .as_deref()
-                                .is_some_and(|tc| tc.contains("?CLASS"))
-                        }) || return_type
-                            .as_deref()
-                            .is_some_and(|rt| rt.contains("?CLASS")))
-                    {
-                        return Err(RuntimeError::typed_msg(
-                            "X::Role::Unimplemented::Multi",
-                            "Unimplemented multi method from role",
-                        ));
-                    }
-                    let resolved_method_name = if let Some(expr) = name_expr {
-                        self.eval_block_value(&[Stmt::Expr(expr.clone())])?
-                            .to_string_value()
-                    } else {
-                        method_name.resolve()
-                    };
-                    let def = MethodDef {
-                        params: params.clone(),
-                        param_defs: param_defs.clone(),
-                        body: std::sync::Arc::new(method_body.clone()),
-                        is_rw: *is_rw,
-                        is_private: *is_private,
-                        is_multi: *multi,
-                        is_my: *is_submethod,
-                        role_origin: None,
-                        original_role: None,
-                        return_type: return_type.clone(),
-                        compiled_code: None,
-                        delegation: None,
-                        is_default: *is_default_candidate,
-                        deprecated_message: None,
-                        is_submethod: *is_submethod,
-                    };
-                    // `my method` in roles are role-private, skip method table.
-                    // Submethods (is_submethod) DO get composed even though
-                    // is_my is true.
-                    let is_role_private = *is_my && !*is_submethod;
-                    if !is_role_private {
-                        if *multi {
-                            role_def
-                                .methods
-                                .entry(resolved_method_name.clone())
-                                .or_default()
-                                .push(def);
-                        } else {
-                            role_def
-                                .methods
-                                .insert(resolved_method_name.clone(), vec![def]);
-                        }
-                    }
-                    // `handles` on a role method: synthesize forwarder methods.
-                    if !is_role_private && !method_handles.is_empty() {
-                        let source_attr_marker = format!("&{}", resolved_method_name);
-                        for spec in method_handles {
-                            match spec {
-                                HandleSpec::Name(target) => {
-                                    role_def
-                                        .methods
-                                        .entry(target.clone())
-                                        .or_default()
-                                        .push(make_delegation_method(&source_attr_marker, target));
-                                }
-                                HandleSpec::Rename { exposed, target } => {
-                                    role_def
-                                        .methods
-                                        .entry(exposed.clone())
-                                        .or_default()
-                                        .push(make_delegation_method(&source_attr_marker, target));
-                                }
-                                HandleSpec::Wildcard => {
-                                    role_def.wildcard_handles.push(source_attr_marker.clone());
-                                }
-                                HandleSpec::Regex(pattern) => {
-                                    role_def
-                                        .wildcard_handles
-                                        .push(format!("{}:regex:{}", source_attr_marker, pattern));
-                                }
-                                HandleSpec::Type(_) => {}
-                            }
-                        }
-                    }
-                }
-                Stmt::Expr(Expr::Call { name, .. })
-                    if name == "__mutsu_stub_die" || name == "__mutsu_stub_warn" =>
-                {
-                    role_def.is_stub_role = true;
-                }
-                Stmt::SetLine(_) => {
-                    // Skip source line annotations
-                }
-                _ => {
-                    if is_parametric {
-                        // Defer non-method/non-attribute statements until composition
-                        // time so they can be re-evaluated with concrete type bindings.
-                        role_def.deferred_body_stmts.push(stmt.clone());
-                    } else {
-                        // Defer execution until after the role is registered so that
-                        // role methods can be called from within the role block body
-                        // (e.g. `role R { method foo {}; R.foo }`).
-                        role_def.deferred_body_stmts.push(stmt.clone());
-                    }
-                }
-            }
-        }
-        // Capture the current environment for anonymous roles so that attribute
-        // defaults referencing closure variables can be evaluated later.
-        let has_expr_default = role_def
-            .attributes
-            .iter()
-            .any(|(_, _, default, ..)| default.is_some());
-        if has_expr_default {
-            role_def.captured_env = Some(self.env.flatten());
-        }
-        // Capture the parents that were added during this registration
-        // (these are the parents specific to this candidate).
-        let candidate_parents = self
-            .registry()
-            .role_parents
-            .get(name)
-            .cloned()
-            .unwrap_or_default();
-        self.registry_mut()
-            .role_candidates
-            .entry(name.to_string())
-            .or_default()
-            .push(RoleCandidateDef {
-                type_params: type_params.to_vec(),
-                type_param_defs: type_param_defs.to_vec(),
-                role_def: role_def.clone(),
-                parents: candidate_parents,
-                language_version: crate::parser::current_language_version(),
-            });
-        if self
-            .registry()
-            .roles
-            .get(name)
-            .is_none_or(|existing| existing.is_stub_role || type_params.is_empty())
-        {
-            self.registry_mut().roles.insert(name.to_string(), role_def);
-            self.registry_mut()
-                .user_declared_roles
-                .insert(name.to_string());
-        }
-        if !type_params.is_empty() && !self.registry().role_type_params.contains_key(name) {
-            self.registry_mut()
-                .role_type_params
-                .insert(name.to_string(), type_params.to_vec());
-        }
-        // When a parametric variant was registered over an existing non-parametric
-        // role (forming a role group), merge the previous parents back into
-        // role_parents so that role_parent_args_for can find all candidates' parents.
-        if let Some(prev) = prev_parents {
-            let mut registry = self.registry_mut();
-            let current = registry.role_parents.entry(name.to_string()).or_default();
-            for p in prev {
-                if !current.contains(&p) {
-                    current.push(p);
-                }
-            }
-        }
-        Ok(())
-    }
-
-    pub(crate) fn register_subset_decl(
-        &mut self,
-        name: &str,
-        base: &str,
-        predicate: Option<&Expr>,
-        version: &str,
-    ) {
-        // When the predicate is `* ~~ <expr>` (Whatever on LHS of SmartMatch),
-        // the parser doesn't wrap it as WhateverCode (to avoid breaking other
-        // smartmatch semantics). Convert it here to a Lambda so the subset
-        // check correctly evaluates `$_ ~~ <expr>` against the candidate value.
-        let predicate = predicate.map(|pred| {
-            if let Expr::Binary {
-                left,
-                op: crate::token_kind::TokenKind::SmartMatch,
-                right,
-            } = pred
-                && matches!(left.as_ref(), Expr::Whatever)
-            {
-                return Expr::Lambda {
-                    param: "_".to_string(),
-                    body: vec![Stmt::Expr(Expr::Binary {
-                        left: Box::new(Expr::Var("_".to_string())),
-                        op: crate::token_kind::TokenKind::SmartMatch,
-                        right: right.clone(),
-                    })],
-                    is_whatever_code: true,
-                };
-            }
-            pred.clone()
-        });
-        // Drop any cached compiled predicate for this name so a redeclaration
-        // recompiles against the new predicate (see `subset_predicate_cache`).
-        self.subset_predicate_cache.remove(name);
-        self.registry_mut().subsets.insert(
-            name.to_string(),
-            SubsetDef {
-                base: base.to_string(),
-                predicate,
-                version: version.to_string(),
-            },
-        );
-        self.env
-            .insert(name.to_string(), Value::Package(Symbol::intern(name)));
-    }
-
-    pub(crate) fn register_cunion_class(&mut self, name: &str) {
-        self.clear_private_zeroarg_method_cache();
-        self.registry_mut().cunion_classes.insert(name.to_string());
-    }
-
-    /// Construct a CUnion instance: all native-int fields share the same
-    /// underlying bytes (little-endian), so setting one field also sets the
-    /// lower bits of the others.
-    pub(crate) fn construct_cunion_instance(
-        &mut self,
-        class_name: &str,
-        args: &[Value],
-    ) -> Result<Value, RuntimeError> {
-        use crate::runtime::native_types::native_int_bounds;
-
-        // Collect class attributes with their types
-        let class_attrs = self.collect_class_attributes(class_name);
-
-        // Parse named args
-        let mut named_args: HashMap<String, Value> = HashMap::new();
-        for arg in args {
-            if let Value::Pair(key, value) = arg {
-                named_args.insert(key.clone(), *value.clone());
-            }
-        }
-
-        // Find the widest provided value and convert to bytes
-        let mut max_bytes = 0u32;
-        let mut raw_value: u64 = 0;
-
-        for (attr_name, _is_public, _default, _is_rw, _, _, _) in &class_attrs {
-            if let Some(val) = named_args.get(attr_name) {
-                let int_val = match val {
-                    Value::Int(i) => *i as u64,
-                    Value::BigInt(n) => {
-                        use num_traits::ToPrimitive;
-                        n.to_u64().unwrap_or(0)
-                    }
-                    _ => 0,
-                };
-                raw_value = int_val;
-                // Find the type of this attribute to determine byte width
-                if let Some(type_constraint) = self.get_attr_type_constraint(class_name, attr_name)
-                {
-                    let byte_width = Self::native_type_byte_width(&type_constraint);
-                    if byte_width > max_bytes {
-                        max_bytes = byte_width;
-                    }
-                }
-            }
-        }
-
-        // Build attributes from shared bytes
-        let bytes = raw_value.to_le_bytes();
-        let mut attrs = HashMap::new();
-        for (attr_name, _is_public, default, _is_rw, _, _, _) in &class_attrs {
-            if let Some(type_constraint) = self.get_attr_type_constraint(class_name, attr_name) {
-                let byte_width = Self::native_type_byte_width(&type_constraint);
-                let val = match byte_width {
-                    1 => Value::Int(bytes[0] as i64),
-                    2 => Value::Int(u16::from_le_bytes([bytes[0], bytes[1]]) as i64),
-                    4 => Value::Int(
-                        u32::from_le_bytes([bytes[0], bytes[1], bytes[2], bytes[3]]) as i64
-                    ),
-                    8 => {
-                        let v = u64::from_le_bytes(bytes);
-                        if let Some((_min, _max)) = native_int_bounds(&type_constraint) {
-                            if type_constraint.starts_with('u') || type_constraint == "byte" {
-                                // Unsigned: store as BigInt if needed
-                                if v > i64::MAX as u64 {
-                                    Value::bigint(num_bigint::BigInt::from(v as u128))
-                                } else {
-                                    Value::Int(v as i64)
-                                }
-                            } else {
-                                // Signed: reinterpret as i64
-                                Value::Int(v as i64)
-                            }
-                        } else {
-                            Value::Int(v as i64)
-                        }
-                    }
-                    _ => {
-                        if let Some(v) = named_args.get(attr_name) {
-                            v.clone()
-                        } else if let Some(expr) = default {
-                            self.eval_block_value(&[Stmt::Expr(expr.clone())])?
-                        } else {
-                            Value::Int(0)
-                        }
-                    }
-                };
-                attrs.insert(attr_name.clone(), val);
-            } else if let Some(v) = named_args.get(attr_name) {
-                attrs.insert(attr_name.clone(), v.clone());
-            } else if let Some(expr) = default {
-                attrs.insert(
-                    attr_name.clone(),
-                    self.eval_block_value(&[Stmt::Expr(expr.clone())])?,
-                );
-            } else {
-                attrs.insert(attr_name.clone(), Value::Int(0));
-            }
-        }
-
-        Ok(Value::make_instance(Symbol::intern(class_name), attrs))
-    }
-
-    /// Get the type constraint for a class attribute, searching MRO.
-    pub(super) fn get_attr_type_constraint(
-        &self,
-        class_name: &str,
-        attr_name: &str,
-    ) -> Option<String> {
-        if let Some(class_def) = self.registry().classes.get(class_name) {
-            for (name, _is_public, _default, _is_rw, _, _, _) in &class_def.attributes {
-                if name == attr_name {
-                    return class_def.attribute_types.get(attr_name).cloned();
-                }
-            }
-        }
-        None
-    }
-
-    fn native_type_byte_width(type_name: &str) -> u32 {
-        match type_name {
-            "uint8" | "int8" | "byte" => 1,
-            "uint16" | "int16" => 2,
-            "uint32" | "int32" => 4,
-            "uint64" | "int64" | "uint" | "int" => 8,
-            _ => 0,
-        }
-    }
-
-    /// Auto-pun a role into a class so it can be instantiated (e.g. via COERCE or new).
-    /// If the role is already registered as a class, this is a no-op.
-    pub(crate) fn ensure_role_punned_to_class(&mut self, role_name: &str) {
-        if self.registry().classes.contains_key(role_name) {
-            return;
-        }
-        self.clear_private_zeroarg_method_cache();
-        let role_def = match self.registry().roles.get(role_name) {
-            Some(r) => r.clone(),
-            None => return,
-        };
-        // Collect attributes and methods from the role itself and all composed parent roles
-        let mut all_attributes = role_def.attributes.clone();
-        let mut all_methods: HashMap<String, Vec<MethodDef>> = role_def.methods.clone();
-        let mut composed_roles_list = vec![role_name.to_string()];
-        if let Some(parent_names) = self.registry().role_parents.get(role_name).cloned() {
-            let mut role_stack: Vec<String> = parent_names;
-            while let Some(parent_role_name) = role_stack.pop() {
-                if !composed_roles_list.contains(&parent_role_name) {
-                    composed_roles_list.push(parent_role_name.clone());
-                    if let Some(parent_role) = self.registry().roles.get(&parent_role_name).cloned()
-                    {
-                        for attr in &parent_role.attributes {
-                            // ClassAttributeDef is a tuple; field 0 is the attribute name
-                            if !all_attributes.iter().any(|a| a.0 == attr.0) {
-                                all_attributes.push(attr.clone());
-                            }
-                        }
-                        for (method_name, method_defs) in &parent_role.methods {
-                            all_methods
-                                .entry(method_name.clone())
-                                .or_default()
-                                .extend(method_defs.clone());
-                        }
-                    }
-                    // Also recurse into grandparent roles
-                    if let Some(grandparents) =
-                        self.registry().role_parents.get(&parent_role_name).cloned()
-                    {
-                        for gp_name in &grandparents {
-                            if !composed_roles_list.contains(gp_name) {
-                                role_stack.push(gp_name.clone());
-                            }
-                        }
-                    }
-                }
-            }
-        }
-        let punned_class = ClassDef {
-            parents: Vec::new(),
-            attributes: all_attributes,
-            attribute_types: HashMap::new(),
-            attribute_smileys: HashMap::new(),
-            attribute_built: HashMap::new(),
-            methods: all_methods,
-            native_methods: HashSet::new(),
-            mro: vec![role_name.to_string(), "Any".to_string(), "Mu".to_string()],
-            wildcard_handles: Vec::new(),
-            alias_attributes: HashSet::new(),
-            class_level_attrs: HashMap::new(),
-        };
-        self.registry_mut()
-            .classes
-            .insert(role_name.to_string(), punned_class);
-        // Register the role and its composed roles
-        self.registry_mut()
-            .class_composed_roles
-            .insert(role_name.to_string(), composed_roles_list);
-        // When punning a bare role (no type params), update the language
-        // revision metadata from the matching candidate so that
-        // `.^language-revision` on the pun instance returns the correct
-        // revision (not the last-registered candidate's revision).
-        let candidate_lang_version =
-            self.registry()
-                .role_candidates
-                .get(role_name)
-                .and_then(|candidates| {
-                    candidates
-                        .iter()
-                        .find(|c| c.type_params.is_empty())
-                        .map(|c| c.language_version.clone())
-                });
-        if let Some(version) = candidate_lang_version {
-            self.store_language_revision_from_version(role_name, &version);
-        }
-    }
-
-    /// Validate that all `$!attr` references in method bodies refer to attributes
-    /// declared directly in the current class (not inherited from a parent).
-    /// In Raku, private attributes are scoped to the declaring class.
-    fn validate_attr_declared_in_class(
-        ctx: &AttrValidationCtx<'_>,
-        stmts: &[Stmt],
-    ) -> Result<(), RuntimeError> {
-        for stmt in stmts {
-            Self::validate_attr_in_stmt(ctx, stmt)?;
-        }
-        Ok(())
-    }
-
-    fn validate_attr_in_stmt(ctx: &AttrValidationCtx<'_>, stmt: &Stmt) -> Result<(), RuntimeError> {
-        match stmt {
-            Stmt::Expr(e) | Stmt::Return(e) | Stmt::Die(e) | Stmt::Fail(e) | Stmt::Take(e, _) => {
-                Self::validate_attr_in_expr(ctx, e)?;
-            }
-            Stmt::VarDecl { expr, .. } => {
-                Self::validate_attr_in_expr(ctx, expr)?;
-            }
-            Stmt::Assign { name, expr, .. } => {
-                // Check if assigning to an undeclared private attribute ($!attr = ...)
-                if let Some(attr_name) = name.strip_prefix('!')
-                    && !attr_name.is_empty()
-                    && !ctx.attrs.contains(attr_name)
-                {
-                    return Err(Self::undeclared_attr_error(ctx, attr_name, "!"));
-                }
-                Self::validate_attr_in_expr(ctx, expr)?;
-            }
-            Stmt::Say(exprs) | Stmt::Put(exprs) | Stmt::Print(exprs) | Stmt::Note(exprs) => {
-                for e in exprs {
-                    Self::validate_attr_in_expr(ctx, e)?;
-                }
-            }
-            Stmt::If {
-                cond,
-                then_branch,
-                else_branch,
-                ..
-            } => {
-                Self::validate_attr_in_expr(ctx, cond)?;
-                Self::validate_attr_declared_in_class(ctx, then_branch)?;
-                Self::validate_attr_declared_in_class(ctx, else_branch)?;
-            }
-            Stmt::While { cond, body, .. } => {
-                Self::validate_attr_in_expr(ctx, cond)?;
-                Self::validate_attr_declared_in_class(ctx, body)?;
-            }
-            Stmt::For { iterable, body, .. } => {
-                Self::validate_attr_in_expr(ctx, iterable)?;
-                Self::validate_attr_declared_in_class(ctx, body)?;
-            }
-            Stmt::Loop {
-                init,
-                cond,
-                step,
-                body,
-                ..
-            } => {
-                if let Some(init) = init.as_ref() {
-                    Self::validate_attr_in_stmt(ctx, init)?;
-                }
-                if let Some(cond) = cond.as_ref() {
-                    Self::validate_attr_in_expr(ctx, cond)?;
-                }
-                if let Some(step) = step.as_ref() {
-                    Self::validate_attr_in_expr(ctx, step)?;
-                }
-                Self::validate_attr_declared_in_class(ctx, body)?;
-            }
-            Stmt::Given { topic, body, .. } => {
-                Self::validate_attr_in_expr(ctx, topic)?;
-                Self::validate_attr_declared_in_class(ctx, body)?;
-            }
-            Stmt::When { cond, body, .. } => {
-                Self::validate_attr_in_expr(ctx, cond)?;
-                Self::validate_attr_declared_in_class(ctx, body)?;
-            }
-            Stmt::Default(body) => {
-                Self::validate_attr_declared_in_class(ctx, body)?;
-            }
-            _ => {}
-        }
-        Ok(())
-    }
-
-    fn undeclared_attr_error(
-        ctx: &AttrValidationCtx<'_>,
-        attr_name: &str,
-        twigil: &str,
-    ) -> RuntimeError {
-        let symbol = format!("${}{}", twigil, attr_name);
-        let message = format!(
-            "Attribute {} not declared in {} {}",
-            symbol, ctx.pkg_kind, ctx.pkg_name
-        );
-        let mut attrs = HashMap::new();
-        attrs.insert("symbol".to_string(), Value::str(symbol.clone()));
-        attrs.insert(
-            "package-name".to_string(),
-            Value::str(ctx.pkg_name.to_string()),
-        );
-        attrs.insert(
-            "package-kind".to_string(),
-            Value::str(ctx.pkg_kind.to_string()),
-        );
-        attrs.insert("what".to_string(), Value::str("attribute".to_string()));
-        attrs.insert("message".to_string(), Value::str(message.clone()));
-        let ex = Value::make_instance(Symbol::intern("X::Attribute::Undeclared"), attrs);
-        let mut err = RuntimeError::new(message);
-        err.exception = Some(Box::new(ex));
-        err
-    }
-
-    fn validate_attr_in_expr(ctx: &AttrValidationCtx<'_>, expr: &Expr) -> Result<(), RuntimeError> {
-        match expr {
-            Expr::Var(name) => {
-                // $! by itself is the error variable, not an attribute
-                if let Some(attr_name) = name.strip_prefix('!')
-                    && !attr_name.is_empty()
-                    && !ctx.attrs.contains(attr_name)
-                {
-                    return Err(Self::undeclared_attr_error(ctx, attr_name, "!"));
-                }
-                // $.attr is compiled as self.attr() — undeclared attributes will
-                // fail at runtime with "No such method", no compile-time check needed.
-            }
-            Expr::MethodCall { target, args, .. } | Expr::HyperMethodCall { target, args, .. } => {
-                Self::validate_attr_in_expr(ctx, target)?;
-                for arg in args {
-                    Self::validate_attr_in_expr(ctx, arg)?;
-                }
-            }
-            Expr::Call { args, .. }
-            | Expr::ArrayLiteral(args)
-            | Expr::BracketArray(args, _)
-            | Expr::StringInterpolation(args) => {
-                for arg in args {
-                    Self::validate_attr_in_expr(ctx, arg)?;
-                }
-            }
-            Expr::Unary { expr, .. }
-            | Expr::PostfixOp { expr, .. }
-            | Expr::Reduction { expr, .. } => {
-                Self::validate_attr_in_expr(ctx, expr)?;
-            }
-            Expr::Binary { left, right, .. }
-            | Expr::MetaOp { left, right, .. }
-            | Expr::HyperOp { left, right, .. } => {
-                Self::validate_attr_in_expr(ctx, left)?;
-                Self::validate_attr_in_expr(ctx, right)?;
-            }
-            Expr::Ternary {
-                cond,
-                then_expr,
-                else_expr,
-            } => {
-                Self::validate_attr_in_expr(ctx, cond)?;
-                Self::validate_attr_in_expr(ctx, then_expr)?;
-                Self::validate_attr_in_expr(ctx, else_expr)?;
-            }
-            Expr::Index { target, index, .. } => {
-                Self::validate_attr_in_expr(ctx, target)?;
-                Self::validate_attr_in_expr(ctx, index)?;
-            }
-            Expr::IndexAssign {
-                target,
-                index,
-                value,
-                ..
-            } => {
-                Self::validate_attr_in_expr(ctx, target)?;
-                Self::validate_attr_in_expr(ctx, index)?;
-                Self::validate_attr_in_expr(ctx, value)?;
-            }
-            Expr::AssignExpr { name, expr, .. } => {
-                // Check if assigning to an undeclared private attribute ($!attr = ...)
-                if let Some(attr_name) = name.strip_prefix('!')
-                    && !attr_name.is_empty()
-                    && !ctx.attrs.contains(attr_name)
-                {
-                    return Err(Self::undeclared_attr_error(ctx, attr_name, "!"));
-                }
-                // $.attr assignment is compiled as self.attr = ... — undeclared
-                // attributes will fail at runtime, no compile-time check needed.
-                Self::validate_attr_in_expr(ctx, expr)?;
-            }
-            Expr::DoBlock { body, .. }
-            | Expr::Block(body)
-            | Expr::Gather(body)
-            | Expr::AnonSub { body, .. }
-            | Expr::AnonSubParams { body, .. }
-            | Expr::Lambda { body, .. } => {
-                Self::validate_attr_declared_in_class(ctx, body)?;
-            }
-            Expr::Try { body: _, catch } => {
-                // Skip attribute validation inside try blocks — accessing an
-                // undeclared attribute will produce a runtime error that the
-                // try block can catch.
-                if let Some(catch) = catch.as_ref() {
-                    Self::validate_attr_declared_in_class(ctx, catch)?;
-                }
-            }
-            Expr::DoStmt(stmt) => {
-                Self::validate_attr_in_stmt(ctx, stmt)?;
-            }
-            _ => {}
-        }
-        Ok(())
-    }
-
-    /// Store a specific language version as type metadata for ^language-revision.
-    pub(crate) fn store_language_revision_from_version(&mut self, name: &str, version: &str) {
-        let revision = language_revision_letter(version);
-        let meta = self.type_metadata.entry(name.to_string()).or_default();
-        meta.insert("language-revision".to_string(), Value::str(revision));
-    }
-}
-
-/// Context threaded through the `$!attr` declaration validators so that an
-/// undeclared private attribute can be reported as a fully-populated
-/// `X::Attribute::Undeclared` (with `package-name`/`package-kind`).
-struct AttrValidationCtx<'a> {
-    attrs: &'a HashSet<String>,
-    pkg_name: &'a str,
-    pkg_kind: &'a str,
-}
-
-/// Extract the language revision letter from a version string like "6.c", "6.d", "6.e".
-fn language_revision_letter(version: &str) -> String {
-    // Version is like "6.c", "6.d", "6.e" — extract the letter after the dot
-    if let Some(letter) = version.strip_prefix("6.") {
-        letter.chars().next().unwrap_or('c').to_string()
-    } else {
-        "c".to_string()
     }
 }
